% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackgroundSubtraction.R
\name{BackgroundSubtraction}
\alias{BackgroundSubtraction}
\title{Working from a completed template, subtracts the respective negative for
each fluorophore}
\usage{
BackgroundSubtraction(data, Normalize = TRUE, keepNegatives = FALSE)
}
\arguments{
\item{data}{The updated template}

\item{Normalize}{Default is TRUE, will normalize the subtracted data so
that scale is from 0 to 100.}

\item{keepNegatives}{Default is FALSE, when TRUE, returns the negatives
used to subtract back to their original rows}
}
\value{
A data.frame with fluorophore signatures
}
\description{
Working from a completed template, subtracts the respective negative for
each fluorophore
}
\examples{
A <- 2 + 2
}
