% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrightnessPlots.R
\name{BrightnessPlots}
\alias{BrightnessPlots}
\title{Visualize raw brightness of single color .fcs files}
\usage{
BrightnessPlots(thedata, input)
}
\arguments{
\item{thedata}{A .csv file containing columns Fluorophore and Detector.}

\item{input}{A path location to where the .fcs files are stored.}
}
\description{
Visualize raw brightness of single color .fcs files
}
\examples{
NULL
}
