% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CartoonSignatures.R
\name{CartoonSignatures}
\alias{CartoonSignatures}
\title{Visualizes fluorophore signatures with smooth density-colored ribbons}
\usage{
CartoonSignatures(
  x = NULL,
  columnname = "Sample",
  data,
  characterColumns = NULL,
  Expand = NULL,
  variance = "medium",
  Normalize = TRUE,
  legend = TRUE,
  plotname = NULL,
  plotlinecolor = NULL,
  ylim = NULL,
  returnType = "Cartoon",
  ribbon_ci = 0.95,
  theme = "viridis",
  bins = 100,
  logScale = TRUE
)
}
\description{
Visualizes fluorophore signatures with smooth density-colored ribbons
}
