% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlData.R
\name{controlData}
\alias{controlData}
\title{Prepare Luciernaga .fcs outputs as Control Inputs for Unmixing}
\usage{
controlData(cs, removestrings)
}
\arguments{
\item{cs}{A flowSet (ex. cs or cs[[1]])}

\item{removestrings}{Portions of the single color names you want removed (Ex. c("DR_ILT_2023_", ".FCS")}
}
\description{
Prepare Luciernaga .fcs outputs as Control Inputs for Unmixing
}
\examples{
NULL
}
