% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosinePlots.R
\name{CosinePlots}
\alias{CosinePlots}
\title{Visualize cosine similarity of raw .fcs files to evaluate single color
controls.}
\usage{
CosinePlots(thedata, input, stats = NULL)
}
\arguments{
\item{thedata}{A data.frame with columns Fluorophore and Detector.}

\item{input}{The location where the .fcs files are stored.}

\item{stats}{Whether to use the median or mean measurement for MFI}
}
\value{
Visualized ggplots for each fluorophore.
}
\description{
Visualize cosine similarity of raw .fcs files to evaluate single color
controls.
}
\examples{
NULL
}
