% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytekQCFilePrep.R
\name{CytekQCFilePrep}
\alias{CytekQCFilePrep}
\title{Converts the Cytek Aurora (TM)'s QC report into a data frame.}
\usage{
CytekQCFilePrep(x, TrackChange)
}
\arguments{
\item{x}{Takes a Levy-Jennings QC tracking report saved as a .csv file, and converts into a "tidyed" dataframe for plotting.
Currently works on our 3L, 4L, 5L Auroras. Please reach out if you find an issue, the .csv export varies a bit and
I want to continue to improve on the code to handle these odd exceptions.}

\item{TrackChange}{Whether to derive a change between previous QC days.}
}
\value{
A dataframe.
}
\description{
Converts the Cytek Aurora (TM)'s QC report into a data frame.
}
\examples{
Not at this time
}
