% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytekQCPlots.R
\name{CytekQCPlots}
\alias{CytekQCPlots}
\title{Takes the dataframe generated from CytekQCFilePrep and returns all the QC plots}
\usage{
CytekQCPlots(x, FailedFlag, pdf, path, filename)
}
\arguments{
\item{x}{The data frame object of Aurora QC data generated by CytekQCFilePrep}

\item{FailedFlag}{Whether to show red flags when a detectors "Out-of-Range" is TRUE}

\item{pdf}{Whether to return a pdf}

\item{path}{Location path to return the pdf}

\item{filename}{File name to save as.}
}
\value{
The pdf and/the plots.
}
\description{
Takes the dataframe generated from CytekQCFilePrep and returns all the QC plots
}
\examples{
NULL
}
