% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluorophoreMatrix.R
\name{FluorophoreMatrix}
\alias{FluorophoreMatrix}
\title{Takes a .csv with Fluorophore and Antigen columns, and generates
a corresponding Cytek Aurora matrix for inter/intra comparisons}
\usage{
FluorophoreMatrix(data, NumberDetectors, returnType = "plot")
}
\arguments{
\item{data}{A path to the .csv or a data.frame object}

\item{NumberDetectors}{The corresponding number of detectors
for the Cytek Aurora}

\item{returnType}{Default is plot}
}
\value{
An assembled plot object for visualization
}
\description{
Takes a .csv with Fluorophore and Antigen columns, and generates
a corresponding Cytek Aurora matrix for inter/intra comparisons
}
\examples{
A <- 2 + 2
}
