% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatePlot.R
\name{GatePlot}
\alias{GatePlot}
\title{Generates called plots from Utility_GatingPlots}
\usage{
GatePlot(x, data, TheDF, gtFile, bins = 270, clearance = 0.2)
}
\arguments{
\item{x}{A specific gate, ex. "nonDebris"}

\item{data}{A GatingSet object}

\item{TheDF}{A data.frame object of the flow file's expr data}

\item{gtFile}{The data.table imported .csv file containing the gating template.}

\item{bins}{Argument to geom_hex for number of bins to visualize the plotted data density.}

\item{clearance}{A buffer area around the plot edge}
}
\value{
A ggplot corresponding to the given inputs
}
\description{
Generates called plots from Utility_GatingPlots
}
\examples{
NULL
}
\keyword{internal}
