% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesis.R
\name{Genesis}
\alias{Genesis}
\title{Internal for LuciernagaQC, creates .fcs files}
\usage{
Genesis(
  x,
  ff,
  minimalfcscutoff = 0.05,
  AggregateName,
  Brightness,
  outpath = NULL,
  OriginalStart,
  OriginalEnd,
  stats = "median",
  NegativeType = "default",
  TotalNegatives = 500,
  Samples = NULL,
  ExportType
)
}
\arguments{
\item{x}{The data.frame of Luciernaga data.}

\item{ff}{An individual cytoset object.}

\item{minimalfcscutoff}{A ratio indicating mininum of the total population needed to split off
into own file, default is set to 0.05}

\item{AggregateName}{Passed final name with modifications from name}

\item{Brightness}{Whether to additionally return a brightness .csv to the outpath}

\item{outpath}{Location to export the fcs and .csv files to}

\item{OriginalStart}{Passed Argument indicating start column for Raw .fcs values}

\item{stats}{Whether "median" or "mean", default is "median"}

\item{NegativeType}{Whether to append a negative pop. Args are "artificial", "internal" and "default"}

\item{TotalNegatives}{How many of the above rows to append, default is set to 500}

\item{Samples}{When Negative type = "Internal", the data.frame of averaged fluorescence per detector}

\item{ExportType}{Passed from above, set to "fcs" for fcs.file return}

\item{OrigingalEnd}{Passed argument indicating end column for raw .fcs values}
}
\description{
Internal for LuciernagaQC, creates .fcs files
}
\keyword{internal}
