% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapPlots.R
\name{HeatmapPlots}
\alias{HeatmapPlots}
\title{Returns a heatmap of subclusters proportion by Fluorophore.}
\usage{
HeatmapPlots(thedata, input, stats = NULL)
}
\arguments{
\item{thedata}{A data.frame with columns Fluorophore and Detector.}

\item{input}{The location where the .fcs files are stored.}

\item{stats}{Whether to use the median or mean measurement for MFI}
}
\value{
Visualized ggplots for each fluorophore.
}
\description{
Returns a heatmap of subclusters proportion by Fluorophore.
}
\examples{
NULL
}
