% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolisticToArchived.R
\name{HolisticToArchived}
\alias{HolisticToArchived}
\title{Converts the extracted BeadData into the ArchivedData format
 needed to plot Gain/RCV fails}
\usage{
HolisticToArchived(
  data,
  manufacturer = "Cytek",
  baselinecutoffs,
  returnTemplate = FALSE,
  outpath = NULL,
  gainmultiplier = 2
)
}
\arguments{
\item{data}{A data.frame object of the bead data}

\item{manufacturer}{Options Cytek or other, basically whether
to use the DailyQC CSV or the template}

\item{baselinecutoffs}{The DailyQC .csv or the template .csv
or associated file.path}

\item{returnTemplate}{Returns DailyQC template that can be
adjusted beyond Cytek settings}

\item{outpath}{Default NULL, specifies where to returnTemplate
.csv to}

\item{gainmultiplier}{Gain times this value is the cutoff point
at which Gain Fails}
}
\value{
An updated data.frame containing the necessary Flag
 columns for plotting
}
\description{
Converts the extracted BeadData into the ArchivedData format
 needed to plot Gain/RCV fails
}
