% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_ParallelGating.R
\name{Internal_ParallelGating}
\alias{Internal_ParallelGating}
\title{Internal for Utility_ParallelNxNPlots}
\usage{
Internal_ParallelGating(
  x,
  x_ff,
  y_ff,
  TheDF,
  yValue,
  columnlist,
  gatelines,
  reference,
  clearance,
  bins,
  AltNameX,
  AltNameY,
  colorX,
  colorY
)
}
\description{
Internal for Utility_ParallelNxNPlots
}
\keyword{internal}
