% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalIterator.R
\name{InternalIterator}
\alias{InternalIterator}
\title{Generates called plots from Utility_GatingPlots}
\usage{
InternalIterator(
  x,
  gs,
  subset,
  gate,
  xValue,
  yValue,
  sample.name,
  removestrings,
  bins
)
}
\arguments{
\item{x}{The name of a individual specimen in a GatingSet}

\item{gs}{The GatingSet Object}

\item{subset}{Desired node of cells to plot}

\item{gate}{Desired gate to show on the same plot as the subset, else NULL}

\item{xValue}{Desired x axis}

\item{yValue}{Desired y axis}

\item{sample.name}{Keyword under which the sample name is stored}

\item{removestrings}{Character values to remove from the name}

\item{bins}{Geo_hex bins for the dots}
}
\value{
A ggplot corresponding to the given inputs
}
\description{
Generates called plots from Utility_GatingPlots
}
\examples{
NULL
}
\keyword{internal}
