% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUnity.R
\name{InternalUnity}
\alias{InternalUnity}
\title{Internal to Utility_UnityPlots}
\usage{
InternalUnity(
  x,
  TheY,
  TheX,
  marginsubset,
  gatesubset,
  sample.name,
  removestrings,
  clearance,
  bins,
  gatelines,
  reference
)
}
\description{
Internal to Utility_UnityPlots
}
\keyword{internal}
