% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Move.R
\name{Luciernag_Move}
\alias{Luciernag_Move}
\title{Transfer the .fcs files selected by LuciernagaTree to a new location}
\usage{
Luciernag_Move(x, data, input, output)
}
\arguments{
\item{x}{A vector containing fluorophore names found within data.}

\item{data}{The data.frame output of LuciernagaTree listing files to be moved.}

\item{input}{The path to the current storage location of the .fcs files.}

\item{output}{The path to the desired future storage location of the selected
files.}
}
\value{
No return, .fcs files are moved desired folder.
}
\description{
Transfer the .fcs files selected by LuciernagaTree to a new location
}
\examples{
NULL
}
