% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Cosine.R
\name{Luciernaga_Cosine}
\alias{Luciernaga_Cosine}
\title{Generates cosine comparison from a data.frame of fluorescent
signatures}
\usage{
Luciernaga_Cosine(
  data,
  returntype = "plot",
  rearrange = TRUE,
  colorlow = "lightblue",
  colorhigh = "red",
  limitlow = 0.4,
  limithigh = 1,
  legend = TRUE
)
}
\arguments{
\item{data}{A data.frame with a single name column and rest
numeric columns}

\item{returntype}{Default returns "plot", alternatively "data"
for underlying data}

\item{rearrange}{Whether to rearrange the plot to group similar
signatures}

\item{colorlow}{Default "lightblue"}

\item{colorhigh}{Default "orange"}

\item{limitlow}{Default 0.4}

\item{limithigh}{Default 1}

\item{legend}{Default TRUE}
}
\value{
Either a ggplot or matrix object
}
\description{
Generates cosine comparison from a data.frame of fluorescent
signatures
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)
library(dplyr)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
UnstainedFCSFiles <- FCS_Files[grep("Unstained", FCS_Files)]
UnstainedCells <- UnstainedFCSFiles[-grep("Beads", UnstainedFCSFiles)]
MyCytoSet <- load_cytoset_from_fcs(UnstainedCells[1],
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <-  c("DR_", "Cells", ".fcs", "-", " ")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

PopulationInterest <- gs_pop_get_data(MyGatingSet[1], subset="lymphocytes")
TheDataValues <- exprs(PopulationInterest[[1]])
TheDataValues <- data.frame(TheDataValues, check.names=FALSE)
Signature <- AveragedSignature(TheDataValues, stats="median")
TheData1 <- Signature[,-grep("Time|FS|SC|SS|Original|W$|H$", names(Signature))]
TheData1 <- TheData1 \%>\% mutate(Sample="lymphocytes") \%>\%
 relocate(Sample, .before=1)

PopulationInterest <- gs_pop_get_data(MyGatingSet[1], subset="nonDebris")
TheDataValues <- exprs(PopulationInterest[[1]])
TheDataValues <- data.frame(TheDataValues, check.names=FALSE)
Signature <- AveragedSignature(TheDataValues, stats="median")
TheData2 <- Signature[,-grep("Time|FS|SC|SS|Original|W$|H$", names(Signature))]
TheData2 <- TheData2 \%>\% mutate(Sample="nonDebris") \%>\%
 relocate(Sample, .before=1)

FinalData <- rbind(TheData1, TheData2)

Plot <- Luciernaga_Cosine(data=FinalData, returntype="plot")

}
