% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_FolderGroupHeatmap.R
\name{Luciernaga_FolderGroupHeatmap}
\alias{Luciernaga_FolderGroupHeatmap}
\title{Small wrapper, passes folder (or files) to Luciernaga_GroupHeatmap}
\usage{
Luciernaga_FolderGroupHeatmap(
  FolderPath,
  cutoff = 0.01,
  removeFollowing = NULL,
  legend = "right",
  returnType = "data"
)
}
\arguments{
\item{FolderPath}{File path to a Folder (or files) to be used
for the heatmap}

\item{cutoff}{Minimal ratio needed for showing, cells not making
criteria rolled into other}

\item{removeFollowing}{Default NULL, string character start for
removal via stringr::str_detect}

\item{legend}{Default is "right", provide "none" to remove.}

\item{returnType}{Default is "data", alternate is "count"}
}
\value{
A GroupHeatmap plot
}
\description{
Small wrapper, passes folder (or files) to Luciernaga_GroupHeatmap
}
\examples{
A <- 2+2
}
