% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_FolderSignatures.R
\name{Luciernaga_FolderSignatures}
\alias{Luciernaga_FolderSignatures}
\title{Takes a folder of Luciernaga_QC signatures and returns a
 data.frame of the signatures contained within, for use
 in downstream data analysis.}
\usage{
Luciernaga_FolderSignatures(
  FolderPath,
  sample.name,
  StringRemoval = NULL,
  fluorophore.name,
  Verbose = FALSE,
  stats = "median",
  PanelCuts = NULL,
  normalize = TRUE,
  returnType = "Signatures"
)
}
\arguments{
\item{FolderPath}{Location of the Luciernaga_QC .fcs outputs}

\item{sample.name}{The keyword where the identifying sample name 
can be found}

\item{StringRemoval}{Default NULL, provide to remove items from
 sample.name
based on values found on the keyword}

\item{fluorophore.name}{Specify the name of the fluorophore,
alternatively NULL}

\item{Verbose}{Default FALSE, provides info as it goes}

\item{stats}{Whether to use median or mean}

\item{PanelCuts}{Default NULL, provide a c(0.5,1) argument to specify 
the brightness percentiles to retrieve signature from for the
individual files}

\item{normalize}{Default TRUE, whether to return normalized or
raw averaged MFI signatures}

\item{returnType}{Default is "Signatures"}
}
\value{
A data.frame of the corresponding averaged signatures per file
}
\description{
Takes a folder of Luciernaga_QC signatures and returns a
 data.frame of the signatures contained within, for use
 in downstream data analysis.
}
\examples{
A <- 2 + 2
}
