% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_SpotCheck.R
\name{Luciernaga_SpotCheck}
\alias{Luciernaga_SpotCheck}
\title{Intended to assess changes at particular detectors within signatures, uses
LuciernagaQC internals with externally provided check at these points}
\usage{
Luciernaga_SpotCheck(
  x,
  subsets,
  sample.name,
  CheckThese,
  inverse.transform = FALSE,
  Increments = 0.1,
  stats = "median"
)
}
\arguments{
\item{x}{A GatingSet object being iterated on}

\item{subsets}{The node for the events of interest}

\item{sample.name}{Keyword variable which samples are stored (ex. "GUID")}

\item{CheckThese}{A vector of the detectors of interest}

\item{inverse.transform}{Default is FALSE}

\item{Increments}{Default is 0.1}

\item{stats}{Default is median}
}
\value{
A LuciernagaQC data style output for detectors of interest
}
\description{
Intended to assess changes at particular detectors within signatures, uses
LuciernagaQC internals with externally provided check at these points
}
\examples{
A <- 2+2
}
