% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Summary.R
\name{Luciernaga_Summary}
\alias{Luciernaga_Summary}
\title{Returns a summary of signature metrics for a gated population}
\usage{
Luciernaga_Summary(
  x,
  y,
  gs,
  externalAF_gs = NULL,
  externalAF_gs_index = NULL,
  externalAF_gs_gate = NULL,
  GuessSimilar = FALSE,
  NumberDetectors = 64,
  excludeThese = "FSC|SSC|Time|-H|-W",
  Unstained = FALSE,
  inverse.transform = TRUE,
  AFOverlap
)
}
\arguments{
\item{x}{Iterated in gate name}

\item{y}{Iterated in specimen index}

\item{gs}{The GatingSet}

\item{externalAF_gs}{TBD}

\item{externalAF_gs_index}{TBD}

\item{externalAF_gs_gate}{TBD}

\item{GuessSimilar}{Default FALSE, will attempt to match to
similar fluorophores in the library}

\item{NumberDetectors}{Default 64 (5-laser Cytek Aurora)}

\item{excludeThese}{Excludes columns containing values, leaving just
 the detector
columns. The default is set to "FSC|SSC|Time|-H|-W"}

\item{Unstained}{Default FALSE, set to TRUE if sample is unstained
 (and subtraction 
therefore is not needed)}

\item{inverse.transform}{Whether to inverse.transform, default is TRUE}

\item{AFOverlap}{See Luciernaga Vignette, default NULL falls back to the 
default shipped within Luciernaga extdata.}
}
\value{
A named list containing the various data.frames and plots for
 the respective gate
}
\description{
Returns a summary of signature metrics for a gated population
}
\examples{
A <- 2 + 2

}
