% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaClustering.R
\name{LuciernagaClustering}
\alias{LuciernagaClustering}
\title{Internal for UnstainedSignatures And SingleStainSignatures}
\usage{
LuciernagaClustering(MyData, These, DetectorName, SkipMain = FALSE)
}
\arguments{
\item{MyData}{The data.frame of raw .fcs exprs values}

\item{These}{A list of columns corresponding to local maxima peaks}

\item{DetectorName}{The brightest peak everything is normalized to.}

\item{SkipMain}{Default FALSE, TRUE used by Luciernaga_SpotCheck}
}
\value{
An internal value
}
\description{
Internal for UnstainedSignatures And SingleStainSignatures
}
\keyword{internal}
