% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaIntegration.R
\name{LuciernagaIntegration}
\alias{LuciernagaIntegration}
\title{Ease-of-life wrapper for Luciernaga_Summary, orchestrates passing of the
respective indices and gates to screen single-color unmixing controls}
\usage{
LuciernagaIntegration(
  template,
  gs,
  AFOverlap = NULL,
  externalAF_gs = NULL,
  externalAF_gs_index = NULL,
  externalAF_gs_gate = NULL,
  excludeThese = "FSC|SSC|Time|-H|-W",
  inverse.transform = TRUE,
  GuessSimilar = FALSE,
  Unstained = FALSE,
  NumberDetectors = 64
)
}
\arguments{
\item{template}{Template designating single-color gates to screen}

\item{gs}{The GatingSet}

\item{AFOverlap}{See Luciernaga Vignette, default NULL falls back to the 
default shipped within Luciernaga extdata.}

\item{externalAF_gs}{Default NULL, if you wish to provide an external AF for
subtraction, provide the gs name (and next two arguments). Leaving NULL will grab
the parent to provided gate on assumption both positive and negative cells will
be present in that gate.}

\item{externalAF_gs_index}{The index value to select the unstained in combination
with previous argument}

\item{externalAF_gs_gate}{The gate corresponding to the unstained population you
want to use for background subtraction}

\item{excludeThese}{Excludes columns containing values, leaving just the detector
columns. The default is set to "FSC|SSC|Time|-H|-W"}

\item{inverse.transform}{Whether to inverse.transform, default is TRUE}

\item{GuessSimilar}{Default FALSE, will attempt to match to
similar fluorophores in the library}

\item{Unstained}{Default FALSE, set to TRUE if sample is unstained (and subtraction 
therefore is not needed)}

\item{NumberDetectors}{Default 64 (5-laser Cytek Aurora)}
}
\value{
A list of lists containing data.frame and plots for all the single-color
reference controls designated in the template.
}
\description{
Ease-of-life wrapper for Luciernaga_Summary, orchestrates passing of the
respective indices and gates to screen single-color unmixing controls
}
\examples{
A <- 2 + 2

}
