% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaLinearSlices.R
\name{LuciernagaLinearSlices}
\alias{LuciernagaLinearSlices}
\title{Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.}
\usage{
LuciernagaLinearSlices(x, subset, sample.name, removestrings, stats, TheReturn)
}
\arguments{
\item{x}{A GatingSet object}

\item{subset}{The desired gating hierarchy level to look at the data}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of values to remove from the name}

\item{stats}{Whether to use "mean" or "median"}

\item{TheReturn}{Whether to return a "raw" or "normalized" value lineplot.}
}
\value{
A ggplot2 object of lineplots from the sliced .fcs file
}
\description{
Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.
}
\examples{
NULL
}
