% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaMove.R
\name{LuciernagaMove}
\alias{LuciernagaMove}
\title{Move the LuciernaTree selected fcs files to a new location}
\usage{
LuciernagaMove(x, data, input, output)
}
\arguments{
\item{x}{A vector of fluorophores in the data}

\item{data}{The data.frame output of LuciernagaTree}

\item{input}{The location of the fcs stored files}

\item{output}{The location to move the selected files}
}
\description{
Move the LuciernaTree selected fcs files to a new location
}
\examples{
NULL
}
