% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaPlot.R
\name{LuciernagaPlot}
\alias{LuciernagaPlot}
\title{Generate Luciernaga plot outputs from List of List}
\usage{
LuciernagaPlot(ListOfList, PlotType, ReturnFolder, CurrentExperiment)
}
\arguments{
\item{ListOfList}{A list containing the list (BrightnessPlot, LinePlot, HeatmapPlot, CosinePlot)}

\item{PlotType}{Whether "html" or "pdf"}

\item{ReturnFolder}{Location to store file}

\item{CurrentExperiment}{Name of Current Experiment}
}
\value{
A file containing the Luciernaga plots in the specified format
}
\description{
Generate Luciernaga plot outputs from List of List
}
\examples{
NULL
}
