% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaReport.R
\name{LuciernagaReport}
\alias{LuciernagaReport}
\title{Converts the Luciernaga outputs into .pdf plots.}
\usage{
LuciernagaReport(
  data,
  RetainedType,
  CellPopRatio,
  outfolder,
  filename,
  LinePlots = TRUE,
  CosinePlots = TRUE,
  StackedBarPlots = TRUE,
  HeatmapPlots = TRUE,
  returntype = "patchwork"
)
}
\arguments{
\item{outfolder}{The location that you want to save the .pdf output to.}

\item{filename}{The name you want to save your .pdf file as.}

\item{returntype}{Passed to Utility_Patchwork for "pdf" or "patchwork"}

\item{LuciernagaData}{A data.frame}

\item{FluorophoreColumnName}{The name of the data.frame column containing your fluorophores.}

\item{ClusterColumnName}{The name of the data.frame column containing your cluster IDs.}
}
\value{
A value to be determined later
}
\description{
Converts the Luciernaga outputs into .pdf plots.
}
\examples{
NULL
}
