% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaReportFromFCS.R
\name{LuciernagaReportFromFCS}
\alias{LuciernagaReportFromFCS}
\title{Visualize cosine similarity of raw .fcs files to evaluate single color
controls.}
\usage{
LuciernagaReportFromFCS(
  path,
  reference,
  stats = "median",
  LinePlots = TRUE,
  CosinePlots = TRUE,
  StackedBarPlots = TRUE,
  Heatmaps = TRUE
)
}
\arguments{
\item{path}{The location to the folder where the Luciernaga .fcs files are stored}

\item{reference}{A path to a .csv file or a dataframe containing Fluorophore and Detector column
information for the panel.}

\item{stats}{Whether to use the median or mean for fluorescent intensity.}

\item{LinePlots}{Return this kind of plot, default is set to TRUE}

\item{CosinePlots}{Return this kind of plot, default is set to TRUE}

\item{StackedBarPlots}{Return this kind of plot, default is set to TRUE}

\item{HeatmapPlots}{Return this kind of plot, default is set to TRUE}
}
\value{
Visualized ggplots for each fluorophore.
}
\description{
Visualize cosine similarity of raw .fcs files to evaluate single color
controls.
}
\examples{
NULL
}
