% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaTree.R
\name{LuciernagaTree}
\alias{LuciernagaTree}
\title{Pick fcs files for unmixing}
\usage{
LuciernagaTree(x, data)
}
\arguments{
\item{x}{A vector of the fluorophores in the data}

\item{data}{The data.frame of the Luciernaga outputs to filter}
}
\description{
Pick fcs files for unmixing
}
\examples{
NULL
}
