% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaUnmix.R
\name{LuciernagaUnmix}
\alias{LuciernagaUnmix}
\title{OLS unmixing of a Gating Set object}
\usage{
LuciernagaUnmix(
  x,
  controlData,
  sample.name,
  addon,
  removestrings,
  subset,
  multiplier,
  outpath
)
}
\arguments{
\item{x}{A Gating Set object}

\item{controlData}{The matrix of single color controls generated from Luciernaga}

\item{sample.name}{The keyword containing the fcs file name}

\item{addon}{Additional addon to append to the new .fcs file name}

\item{removestrings}{A list of values to remove from name}

\item{subset}{A gating hierarchy level to sort cells at, expression values retrieved from these}

\item{multiplier}{A number to scale the OLS coefficients by}

\item{outpath}{The return folder for the .fcs files}
}
\value{
A new .fcs file with the new columns appended
}
\description{
OLS unmixing of a Gating Set object
}
\examples{
NULL
}
