% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameForSample.R
\name{NameForSample}
\alias{NameForSample}
\title{Generate a name from the sample for plot identification}
\usage{
NameForSample(
  x,
  sample.name,
  removestrings,
  experiment.name = NULL,
  experiment = NULL,
  condition.name = NULL,
  condition = NULL
)
}
\arguments{
\item{x}{A GatingSet object}

\item{sample.name}{The .fcs keyword that contains an unique name for that sample}

\item{removestrings}{A string of character values to remove from the sample.name}

\item{experiment.name}{The .fcs keyword that contains the name of the experiment}

\item{experiment}{Directly provide the name of the experiment (alternative to experiment.name)}

\item{condition.name}{The .fcs keyword that contains the name of the condition.}

\item{condition}{Directly provide the name of the condition (alternative to condition.name)}
}
\value{
A name matching the provided information for use in plot labeling.
}
\description{
Generate a name from the sample for plot identification
}
\examples{
NULL
}
\keyword{internal}
