% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniversalIterator.R
\name{ParallelUniversalIterator}
\alias{ParallelUniversalIterator}
\title{Internal for Utility_ParallelNbyNPlots}
\usage{
ParallelUniversalIterator(
  x,
  x_ff,
  y_ff,
  TheDF,
  yValue,
  columnlist,
  gatelines,
  reference,
  clearance,
  bins,
  AltNameX,
  AltNameY,
  colorX,
  colorY
)
}
\description{
Internal for Utility_ParallelNbyNPlots
}
\keyword{internal}
