% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_ChorusPDF.R
\name{QC_ChorusPDF}
\alias{QC_ChorusPDF}
\title{Processes BD Chorus QC PDF files into .csv files}
\usage{
QC_ChorusPDF(
  x,
  returnPreference = "Imaging",
  returnType = "data",
  outpath = NULL
)
}
\arguments{
\item{x}{A file.path to the desired QC pdf}

\item{returnPreference}{When both modes present, whether
to return QC values from Imaging or High-Speed settings}

\item{returnType}{Default data, alternative is csv}

\item{outpath}{When return type is csv, file.path to store the csv}
}
\description{
Processes BD Chorus QC PDF files into .csv files
}
\examples{
A <- 2+2

}
