% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Plots.R
\name{QC_Plots}
\alias{QC_Plots}
\title{Takes the dataframe generated from CytekQCFilePrep and returns all the QC
plots}
\usage{
QC_Plots(
  x,
  FailedFlag,
  MeasurementType = NULL,
  Metadata = NULL,
  plotType = "individual",
  returntype,
  path,
  filename,
  thecolumns = 1,
  therows = 3,
  width = 7,
  height = 9,
  strict = FALSE,
  YAxisLabel = NULL,
  RepairVisits = NULL
)
}
\arguments{
\item{x}{The data frame object of Aurora QC data generated by
CytekQCFilePrep}

\item{FailedFlag}{Whether to show red flags when a detectors
"Out-of-Range" is TRUE}

\item{MeasurementType}{A list containing character values, will select
any present from the column names for plotting}

\item{Metadata}{An optional character column to use as ggplot factor.}

\item{plotType}{Whether to return "individual" plots or "comparison"
plots (provide Metadata name!)}

\item{returntype}{Whether to return to pdf, patchwork or plots.}

\item{path}{Location path to return the pdf}

\item{filename}{File name to save as.}

\item{thecolumns}{The number of columns per page}

\item{therows}{The number of rows per page}

\item{width}{Desired page width}

\item{height}{Desired page height}

\item{strict}{Default FALSE, when TRUE, parameters must be exact match
for MeasurementType rather than simply containing those character
strings.}

\item{YAxisLabel}{Character string for the y-axis label.}

\item{RepairVisits}{Passed data.frame of engineer visits for vertical
lines, default NULL.}
}
\value{
The pdf and/the plots.
}
\description{
Takes the dataframe generated from CytekQCFilePrep and returns all the QC
plots
}
\examples{
library(ggplot2)

File_Location <- system.file("extdata", package = "Luciernaga")
CSV_Pattern <- ".CSV$"
CSV_Files <- list.files(path = File_Location, pattern = CSV_Pattern,
                       full.names = TRUE)
TidyData <- QC_FilePrep(CSV_Files, TrackChange = FALSE)

StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

SinglePlot <- "UV7-Gain"

TheSinglePlot <- QC_Plots(x = TidyData, MeasurementType = SinglePlot,
                         FailedFlag = TRUE, returntype = "patchwork",
                         path = StorageLocation,
                         filename = "CytekAurora5L_QC")
}
