% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_ProspectiveAdditions.R
\name{QC_ProspectiveAdditions}
\alias{QC_ProspectiveAdditions}
\title{From existing panel, figures out open detectors, and returns potential
 fluorophores that might fit between the existing ones}
\usage{
QC_ProspectiveAdditions(
  path,
  NumberDetectors,
  TheCutoff = 0.9,
  returnAll = FALSE,
  filename,
  outpath,
  returnCSV
)
}
\arguments{
\item{path}{Filepath to the panel .csv, or a data.frame with column
Fluorophore}

\item{NumberDetectors}{Number of detectors for the instrument to pull
references}

\item{TheCutoff}{Default is 0.9, determines HighOverlap count threshold}

\item{returnAll}{Whether to return all variants, default is FALSE}

\item{filename}{Desired name for the output .csv}

\item{outpath}{Path Location to store the the output .csv}

\item{returnCSV}{Whether to return as a csv to designated outpath and
filename, a TRUE/FALSE}
}
\value{
A csv containing selected fluorophores
}
\description{
From existing panel, figures out open detectors, and returns potential
 fluorophores that might fit between the existing ones
}
\examples{
Folder_Location <- system.file("extdata", package = "Luciernaga")
ThePanelLocation <- list.files(Folder_Location, pattern = "^Panel.csv",
 full.names = TRUE)
OutPath <- file.path("C:", "Users", "JohnDoe", "Desktop")

ProspectiveAdditions <- QC_ProspectiveAdditions(path = ThePanelLocation,
NumberDetectors = 64, TheCutoff = 0.9, returnAll = FALSE,
returnCSV = FALSE, filename = "ProspectiveAdditions", outpath = OutPath)

}
