% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_ViewSignature.R
\name{QC_ViewSignature}
\alias{QC_ViewSignature}
\title{Visualizes the Signature for given row in an averaged signature data.frame.}
\usage{
QC_ViewSignature(x, data, Normalize = TRUE)
}
\arguments{
\item{x}{Name in the Sample column you want to filter for}

\item{data}{A data.frame object from QC_LibraryParse containing Fluorophore name column and numeric detector columns.}

\item{Normalize}{Whether to normalize the data based on peak detector value, default is TRUE}
}
\value{
A dataframe of similar fluorophores
}
\description{
Visualizes the Signature for given row in an averaged signature data.frame.
}
\examples{
NULL
}
