% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_WhatsThis.R
\name{QC_WhatsThis}
\alias{QC_WhatsThis}
\title{Querries reference signatures and returns most similar fluorophores by cosine similarity.}
\usage{
QC_WhatsThis(x, data, NumberHits)
}
\arguments{
\item{x}{Name from Sample Column}

\item{data}{The dataframe object from QC_LibraryParse}

\item{NumberHits}{Number of most similar fluorophores by cosine.}
}
\value{
A dataframe of similar fluorophores
}
\description{
Querries reference signatures and returns most similar fluorophores by cosine similarity.
}
\examples{
NULL
}
