% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Rest.R
\name{RestTable}
\alias{RestTable}
\title{Internal for Wetlab_Rest}
\usage{
RestTable(data, outpath = NULL, filename = "CellResuspensions")
}
\arguments{
\item{data}{The resuspension data from WetlabRest}

\item{outpath}{The desired storage location for the .png file}

\item{filename}{The desired name of the .png file}
}
\description{
Internal for Wetlab_Rest
}
