% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_RidgePlot.R
\name{RidgePlots}
\alias{RidgePlots}
\title{Internal to Utility_RidgePlots}
\usage{
RidgePlots(x, cs, TheY, TheFill)
}
\arguments{
\item{x}{Passed Argument for X axis}

\item{cs}{The CytoSet with the data}

\item{TheY}{The passed argument for the Y axis}

\item{TheFill}{The passed argument for the Fill color by factor.}
}
\description{
Internal to Utility_RidgePlots
}
\keyword{internal}
