% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleColorPlots.R
\name{SingleColorPlots}
\alias{SingleColorPlots}
\title{Plot the Gating Hierarchy to make sure the gating was appropiate}
\usage{
SingleColorPlots(
  x,
  subsets,
  column,
  sample.name,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  bins,
  clearance,
  outpath,
  sourcelocation
)
}
\arguments{
\item{x}{A Gating Set object (example, gs or gs[[1]])}

\item{subsets}{The Gating Hierarchy node of cells to be plotted (ex. "root")}

\item{column}{Unclear}

\item{sample.name}{Keyword under which samples names are stored (ex. "GUID")}

\item{experiment}{Provide an experiment name directly. (ex. "Jan2024")}

\item{experiment.name}{Keyword under which experiment names are stored (ex. "GROUPNAME")}

\item{condition}{Provide a condition name directly (ex. "PMA")}

\item{condition.name}{Keyword under which condition names are stored (ex. "TUBENAME")}

\item{bins}{Bins for which the ggplot object (ex. 270)}

\item{clearance}{Margin by which the quantile MFI is multiplied by to avoid clipping main pop, but exclude outliers}

\item{outpath}{Location where the generated .pdf should be stored.}

\item{sourcelocation}{Location where the .csv for openCyto gating is stored.}
}
\description{
Plot the Gating Hierarchy to make sure the gating was appropiate
}
\examples{
NULL
}
