% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleStainSignatures.R
\name{SingleStainSignatures}
\alias{SingleStainSignatures}
\title{Internal to LuciernagaQC}
\usage{
SingleStainSignatures(
  x,
  WorkAround1,
  ColsN,
  StartNormalizedMergedCol,
  EndNormalizedMergedCol,
  Samples,
  Increments = 0.1,
  Subtraction = "Internal",
  stats,
  ratioSCcutoff = 0.01,
  TheMainAF,
  AggregateName,
  Verbose = FALSE,
  SCData = "subtracted",
  LocalMaximaRatio,
  SecondaryPeaks
)
}
\arguments{
\item{x}{The Detector being mapped in}

\item{WorkAround1}{The data.frame being used}

\item{ColsN}{Indicated end of Raw Value Columns}

\item{StartNormalizedMergedCol}{Indicated Start Normalized Columns}

\item{EndNormalizedMergedCol}{Indicated End Normalized Columns}

\item{Samples}{A data.frame row containing raw data of external Autofluorescence}

\item{Increments}{A numeric to round the normalized bins by. Default is 0.1}

\item{stats}{A passed param, usually "median" or "mean"}

\item{ratioSCcutoff}{The ratio dictating mininum size return for single colors, default is 0.01}

\item{TheMainAF}{A passed parameter dictating what detector column the autofluorescence is}

\item{AggregateName}{A passed parameter deriving from edits to sample.name and removestrings}

\item{Verbose}{Whether to return intermediate objects via print and plot for progress monitoring}

\item{SCData}{Whether to return "subtracted" or "raw" data at the end.}

\item{LocalMaximaRatio}{Height of peaks to proceed}

\item{SecondaryPeaks}{Number of Secondary Peaks, default is set to 2.}

\item{Substraction}{Dictates what autofluorescence gets extracted. Options include "Internal", "Average"}
}
\description{
Internal to LuciernagaQC
}
\keyword{internal}
