% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexReduction.R
\name{StainBrightnessIndexCalculator}
\alias{StainBrightnessIndexCalculator}
\title{Takes the folder outputs from , parses the stain index for all the combinations,
returning as a large long data.frame}
\usage{
StainBrightnessIndexCalculator(
  folder_location,
  outpath = NULL,
  excludeThese = "FSC|SSC|Time",
  channelRange = 4096,
  maxValue = 4194304,
  pos = 5.62,
  neg = 0,
  widthBasis = -1000,
  inverse.transform = TRUE,
  stringAppend = "-A"
)
}
\arguments{
\item{folder_location}{File.path to the parent folder where all the subfolders containing
the variant .fcs files were stored}

\item{outpath}{Where you want to store the processed data}

\item{excludeThese}{Used to exclude columns from transformation, default is "FSC|SSC|Time"}

\item{channelRange}{Default for biexponential transformation is 4096}

\item{maxValue}{Default for biexponential transformation is 4194304}

\item{pos}{Default for biexponential transformation is 5.62}

\item{neg}{Default for biexponential transformation is 0}

\item{widthBasis}{Default for biexponential transformation is -1000}

\item{inverse.transform}{Default is TRUE}

\item{stringAppend}{Default is -A}
}
\value{
A data.frame containing staining index for all the FMO folders fcs files.
}
\description{
Takes the folder outputs from , parses the stain index for all the combinations,
returning as a large long data.frame
}
\examples{
A <- 2+2


}
