% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexReduction.R
\name{StainIndexLocal}
\alias{StainIndexLocal}
\title{Internal for BrightnessIndexIterator (also an internal). When handed the gated flowframe,
proceeds to extract the correct positive and negative gate values for the respective single-color
and process the staining index.}
\usage{
StainIndexLocal(x, inverse.transform)
}
\arguments{
\item{x}{The individual GatingHierarchy being iterated in}

\item{inverse.transform}{Whether to reverse the transformation 
before calculating the staining index.}
}
\description{
Internal for BrightnessIndexIterator (also an internal). When handed the gated flowframe,
proceeds to extract the correct positive and negative gate values for the respective single-color
and process the staining index.
}
