% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClusters.R
\name{subClusters}
\alias{subClusters}
\title{Check identical clusters for individual variability across experiments.}
\usage{
subClusters(data)
}
\arguments{
\item{data}{The nested portion of the data.frame, containing the detectors
and the experiment column}
}
\value{
A variable to be defined later
}
\description{
Check identical clusters for individual variability across experiments.
}
\examples{
NULL
}
