% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniversalIterator.R
\name{UniversalIterator}
\alias{UniversalIterator}
\title{Internal for Utility_NbyNPlots}
\usage{
UniversalIterator(
  x,
  x_ff,
  TheDF,
  yValue,
  columnlist,
  gatelines,
  reference,
  clearance,
  bins,
  AltNameX,
  AltNameY,
  colorX,
  colorY
)
}
\description{
Internal for Utility_NbyNPlots
}
\keyword{internal}
