% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_ff.R
\name{unmix_ff}
\alias{unmix_ff}
\title{Unmix your .fcs samples.}
\usage{
unmix_ff(fs, control, unmixMethod, multiplier, outpath)
}
\arguments{
\item{fs}{A flowSet object (ex. fs or fs[[1]])}

\item{control}{The output of controlData}

\item{unmixMethod}{Choice of unmixing method}

\item{multiplier}{Value unmixed values are multiplied by for better visualization}

\item{outpath}{Location the unmixed .fcs files are returned to.}
}
\description{
Unmix your .fcs samples.
}
\examples{
NULL
}
