% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnstainedSignatures.R
\name{UnstainedSignatures}
\alias{UnstainedSignatures}
\title{Internal to LuciernagaQC}
\usage{
UnstainedSignatures(
  x,
  WorkAround1,
  alternatename,
  ColsN,
  StartNormalizedMergedCol,
  EndNormalizedMergedCol,
  Increments,
  Verbose = FALSE,
  LocalMaximaRatio = 0.15,
  SecondaryPeaks
)
}
\arguments{
\item{x}{A passed detector used for filtering}

\item{WorkAround1}{A passed data.frame}

\item{alternatename}{Alternate name descriptor}

\item{ColsN}{Indicated end of Raw Value Columns}

\item{StartNormalizedMergedCol}{Indicated Start Normalized Columns}

\item{EndNormalizedMergedCol}{Indicated End Normalized Columns}

\item{Increments}{A numeric to round the normalized bins by. Default is 0.1}

\item{Verbose}{Whether to return intermediate objects via print and plot for
progress monitoring}

\item{LocalMaximaRatio}{Height of peaks to proceed}

\item{SecondaryPeaks}{Number of Secondary Peaks, default is set to 2.}
}
\value{
An internal value
}
\description{
Internal to LuciernagaQC
}
\keyword{internal}
