% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Autofluorescence.R
\name{Utility_Autofluorescence}
\alias{Utility_Autofluorescence}
\title{Runs Luciernaga Comparison}
\usage{
Utility_Autofluorescence(
  x,
  sample.name,
  removestrings,
  subsets,
  columns,
  subsample,
  outpath
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subsets}{The subset of interest from gating hierarchy}

\item{columns}{A subset of columns to pass instead}

\item{subsample}{If downsampling is wanted.}

\item{outpath}{Location to store new .fcs files}
}
\value{
Numbers
}
\description{
Runs Luciernaga Comparison
}
\examples{
Not applicable
}
