% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_DensityOverlay.R
\name{Utility_DensityOverlay}
\alias{Utility_DensityOverlay}
\title{Generates an overlay plot containing all objects in the GatingSet}
\usage{
Utility_DensityOverlay(
  gs,
  subset,
  inverse.transform = FALSE,
  TheX = NULL,
  TheFill,
  returntype,
  outpath,
  filename,
  therows = 3,
  thecolumns = 3,
  width = 7,
  height = 9
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{subset}{The desired Gating Hierarchy node (ex. "lymphocytes")}

\item{inverse.transform}{Default is FALSE, reverses transformation applied
to GatingSet as its converted to a CytoSet}

\item{TheX}{A desired marker to plot, leave NULL for all markers}

\item{TheFill}{A desired marker to color individual specimens by (as named
in pData)}

\item{returntype}{Whether to return a "pdf", "patchwork" or "plots"}

\item{outpath}{Desired storage location}

\item{filename}{The file name for the new .pdf}

\item{therows}{The desired number of rows per page}

\item{thecolumns}{The desired number of columns per page}

\item{width}{The desired page width}

\item{height}{The desired page height}
}
\value{
A pdf file or ggplot object
}
\description{
Generates an overlay plot containing all objects in the GatingSet
}
\examples{

library(BiocGenerics)
library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(
  path = File_Location, pattern = ".fcs",
  full.names = TRUE
)
Unmixed_FullStained <- FCS_Files[grep("Unmixed", FCS_Files)]
UnmixedFCSFiles <- Unmixed_FullStained[1:2]
UnmixedCytoSet <- load_cytoset_from_fcs(
  UnmixedFCSFiles[1:2],
  truncate_max_range = FALSE, transformation = FALSE
)
UnmixedGatingSet <- GatingSet(UnmixedCytoSet)
Markers <- colnames(UnmixedCytoSet)
KeptMarkers <- Markers[-grep("Time|FS|SC|SS|Original|-W$|-H$|AF", Markers)]
MyBiexponentialTransform <- flowjo_biexp_trans(
  channelRange = 256,
  maxValue = 1000000, pos = 4.5, neg = 0, widthBasis = -1000
)
TransformList <- transformerList(KeptMarkers, MyBiexponentialTransform)
UnmixedGatingSet <- flowWorkspace::transform(
  UnmixedGatingSet, TransformList
)
FileLocation <- system.file("extdata", package = "Luciernaga")
UnmixedGates <- fread(
  file.path(path = FileLocation, pattern = "GatesUnmixed.csv")
)
UnmixedGating <- gatingTemplate(UnmixedGates)
gt_gating(UnmixedGating, UnmixedGatingSet)

removestrings <- c("DTR_", ".fcs")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

Condition <- data.frame(Condition = c("Ctrl", "Ctrl"))
pd <- pData(UnmixedGatingSet)
new_pd <- cbind(pd, Condition)
pData(UnmixedGatingSet) <- new_pd

Plot <- Utility_DensityOverlay(
  gs = UnmixedGatingSet, subset = "lymphocytes",
  TheX = "APC-Fire 810-A", TheFill = "Condition", returntype = "plots",
  outpath = "C:/Users/JohnDoe/Desktop/", filename = "CD38_Expression"
)

}
