% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Fill.R
\name{Utility_Fill}
\alias{Utility_Fill}
\title{Internal function for relative brightness}
\usage{
Utility_Fill(x)
}
\arguments{
\item{x}{The data.frame object passed by relative brightness}
}
\description{
Internal function for relative brightness
}
\examples{
NULL
}
