% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralNbyNGating.R
\name{Utility_GeneralGating}
\alias{Utility_GeneralGating}
\title{Generate ggplot2s for different channels}
\usage{
Utility_GeneralGating(
  x,
  ff,
  yValue,
  clearance,
  bins,
  columnlist,
  TheDF,
  gatelines,
  reference = NULL
)
}
\arguments{
\item{x}{Passed channel}

\item{ff}{The dataframe data for that sample}

\item{yValue}{What wanted on the yaxis}

\item{columnlist}{list all channels with x removed}

\item{TheDF}{the external limits settings}

\item{gatelines}{whether to plot the ModernCutoffLines from reference dataframe}

\item{reference}{location ModernCutoff dataframe.}
}
\description{
Generate ggplot2s for different channels
}
\examples{
NULL
}
