% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_LinearSlices.R
\name{Utility_LinearSlices}
\alias{Utility_LinearSlices}
\title{Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.}
\usage{
Utility_LinearSlices(x, subset, sample.name, removestrings, TheReturn)
}
\arguments{
\item{x}{A GatingSet object}

\item{subset}{The desired gating hierarchy level to look at the data}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of values to remove from the name}

\item{TheReturn}{Whether to return a "raw" or "normalized" value lineplot.}
}
\value{
A ggplot2 object of lineplots from the sliced .fcs file
}
\description{
Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.
}
\examples{
Not at this time
}
