% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Luciernaga.R
\name{Utility_Luciernaga}
\alias{Utility_Luciernaga}
\title{Utility_Luciernaga}
\usage{
Utility_Luciernaga(
  LuciernagaData,
  LuciernagaVariables,
  LuciernagaClusters,
  outfolder,
  filename
)
}
\arguments{
\item{LuciernagaData}{A data.frame}

\item{LuciernagaVariables}{The data.frame column that contains your flourophores (ex. "Samples" -> "APCFire750")}

\item{LuciernagaClusters}{The data.frame column that contains your clusters (ex. "Clusters" -> "R01_10-R02_10-R03_10")}

\item{outfolder}{The location that you want to save the .pdf output to.}

\item{filename}{The name you want to save your .pdf file as.}
}
\description{
Utility_Luciernaga
}
\examples{
NULL

}
