% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_MonocyteZombies.R
\name{Utility_MonocyteZombies}
\alias{Utility_MonocyteZombies}
\title{Adds additional gates for monocytes, dead cells, or their respective unstained negatives}
\usage{
Utility_MonocyteZombies(x, sample.name, bins)
}
\arguments{
\item{x}{A Gating Set object (ex. gs or gs[[1]])}

\item{sample.name}{Keyword for which sample name is stored}

\item{bins}{Number of bins resulting ggplot2 object should be plotted with}
}
\description{
Adds additional gates for monocytes, dead cells, or their respective unstained negatives
}
\examples{
NULL
}
