% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParallelNxNPlots.R
\name{Utility_ParallelNbyNPlots}
\alias{Utility_ParallelNbyNPlots}
\title{Overlay plot two .fcs files from two different gs.}
\usage{
Utility_ParallelNbyNPlots(
  x,
  y,
  sample.name,
  removestrings,
  Override = FALSE,
  marginsubset,
  gatesubset,
  ycolumn,
  bins,
  clearance,
  colorX,
  colorY,
  gatelines,
  reference = NULL,
  outpath,
  pdf
)
}
\arguments{
\item{x}{The first gs object}

\item{y}{The second gs object}

\item{sample.name}{The keyword under which sample names are stored}

\item{removestrings}{A list of values to remove from sample names}

\item{Override}{Exclude raw columns with -A$}

\item{marginsubset}{The gs subset that defines the plot margin}

\item{gatesubset}{The gs subset of interest}

\item{ycolumn}{The desired y-column for the comparisons}

\item{bins}{Desired number of hex bins}

\item{clearance}{A multiplication factor for margin wiggle room (0.2)}

\item{colorX}{Color for the x gs}

\item{colorY}{Color for the y gs}

\item{gatelines}{Whether to plot .csv specified gate lines}

\item{reference}{Reference for .csv specified gate lines}

\item{outpath}{Location which to store the output}

\item{pdf}{Whether to return as a pdf}
}
\value{
Either list ggplot objects or a pdf object
}
\description{
Overlay plot two .fcs files from two different gs.
}
\examples{

library(BiocGenerics)
library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
Unmixed_FullStained <- FCS_Files[grep("Unmixed", FCS_Files)]
UnmixedFCSFiles <- Unmixed_FullStained[1:2]
UnmixedCytoSet <- load_cytoset_from_fcs(UnmixedFCSFiles[1:2],
  truncate_max_range = FALSE,transformation = FALSE)
UnmixedGatingSet <- GatingSet(UnmixedCytoSet)
Markers <- colnames(UnmixedCytoSet)
KeptMarkers <- Markers[-grep("Time|FS|SC|SS|Original|-W$|-H$|AF", Markers)]
MyBiexponentialTransform <- flowjo_biexp_trans(channelRange = 256,
  maxValue = 1000000,pos = 4.5, neg = 0, widthBasis = -1000)
TransformList <- transformerList(KeptMarkers, MyBiexponentialTransform)
UnmixedGatingSet <- flowWorkspace::transform(UnmixedGatingSet, TransformList)
FileLocation <- system.file("extdata", package = "Luciernaga")
UnmixedGates <- fread(file.path(path = FileLocation, pattern = 'GatesUnmixed.csv'))
UnmixedGating <- gatingTemplate(UnmixedGates)
gt_gating(UnmixedGating, UnmixedGatingSet)

removestrings <-  c("DTR_", ".fcs")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

Plot <- Utility_ParallelNbyNPlots(x=UnmixedGatingSet[1], y = UnmixedGatingSet[2],
 sample.name = "GROUPNAME", removestrings = ".fcs", Override = FALSE,
 marginsubset = "lymphocytes", gatesubset = "live", ycolumn = "Spark Blue 550-A",
  bins = 120, clearance = 0.2, colorX = "lightblue", colorY = "orange", gatelines = FALSE,
  reference = NULL, outpath = StorageLocation,pdf = FALSE)

}
