% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_RelativeBrightness.R
\name{Utility_RelativeBrightness}
\alias{Utility_RelativeBrightness}
\title{Helper function splits Cluster into individual components}
\usage{
Utility_RelativeBrightness(x)
}
\arguments{
\item{x}{A data.frame containing column Cluster}
}
\description{
Helper function splits Cluster into individual components
}
\examples{
NULL
}
