% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_SingleColorQC.R
\name{Utility_SingleColorQC}
\alias{Utility_SingleColorQC}
\title{Main Luciernaga Function, normalized on individual cell level}
\usage{
Utility_SingleColorQC(
  x,
  subsets,
  sample.name,
  removestrings,
  Verbose = FALSE,
  unmixingcontroltype = "both",
  Unstained = FALSE,
  ratiopopcutoff = 0.01,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  AFOverlap,
  stats = "median",
  desiredAF = NULL,
  externalAF = NULL,
  ExportType,
  SignatureReturnNow = FALSE,
  sourcelocation,
  outpath,
  artificial,
  Brightness = FALSE
)
}
\arguments{
\item{x}{A Gating Set object (ex. gs or gs[[1]])}

\item{subsets}{The Gating Hierarchy level you will be sampling at}

\item{sample.name}{Keyword variable which samples are stored (ex. "GUID")}

\item{removestrings}{A string of character values to remove from sample.name}

\item{Verbose}{Whether to print outputs as you go.}

\item{unmixingcontroltype}{Whether your inputs are "cells", "beads" or "both"}

\item{ratiopopcutoff}{A numeric ratio for peak detector inclusion, default is set to 0.01 all startingcells}

\item{experiment}{Provide directly experiment name (ex. "JAN2024")}

\item{experiment.name}{Keyword variable which experiment information
is stored (ex. "TUBENAME")}

\item{condition}{Provide directly experiment name (ex. "JAN2024")}

\item{condition.name}{Keyword variable which experiment information
is stored (ex. "TUBENAME")}

\item{AFOverlap}{A data.frame or a filepath to the CSV containing the Autofluorescence
overlap of individual fluorophores for exclusion}

\item{stats}{Whether to take "mean" or "median"}

\item{desiredAF}{Main Autofluorescence Detector (ex. "V7-A")}

\item{externalAF}{A data.frame row containing external autofluorescence to subtract from single colors.}

\item{ExportType}{Whether to return "fcs", "data.frame" or "csv"}

\item{SignatureReturnNow}{Short circuits the function and returns signature for specified autofluorescence.}

\item{sourcelocation}{Location where .fcs creation file is stored}

\item{outpath}{Location where created .fcs and .csv files are sent}

\item{artificial}{Whether an artificial 0 population should be added for
a background autofluorescence stand in.}

\item{Brightness}{Whether sum of detectors should be returned.}

\item{Unmixing}{Set to True when running unstained samples that don't have Unstained in the Name}
}
\value{
Additional information to be added
}
\description{
Main Luciernaga Function, normalized on individual cell level
}
\examples{
NULL
}
