% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_SingleColors.R
\name{Utility_SingleColors}
\alias{Utility_SingleColors}
\title{Plot Individual Fluorophore Gates, take the MFI, return the values.}
\usage{
Utility_SingleColors(
  x,
  sample.name,
  experiment = NULL,
  experiment.date,
  rootlevel,
  samplinglevel,
  bins,
  stats = NULL,
  outpath,
  source,
  sourcelocation
)
}
\arguments{
\item{x}{A GatingSet type object (ex. gs or gs[[1]])}

\item{sample.name}{Keyword variable under which samples indentity is stored (ex. "GUID")}

\item{experiment}{Provide directly the experiment label (ex. "Jan2024")}

\item{experiment.date}{Keyword variable under which experiment is stored in absence of Experiment}

\item{rootlevel}{Gating Hierarchy node under which new gates will be constructed}

\item{samplinglevel}{Gating Hierarchy level from which MFI will be taken}

\item{bins}{When plotting, how many bins resolution}

\item{stats}{Whether to use "mean" or "median"}

\item{outpath}{Location which to save output}

\item{source}{Whether to export for plotting}

\item{sourcelocation}{Location where source plotting file is stored.}
}
\description{
Plot Individual Fluorophore Gates, take the MFI, return the values.
}
\examples{
NULL
}
