% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Subtract.R
\name{Utility_Subtract}
\alias{Utility_Subtract}
\title{Originial Subtract Autofluorescence from Median Single Colors}
\usage{
Utility_Subtract(i, Data)
}
\arguments{
\item{i}{A row of the data.frame object}

\item{Data}{A data.frame object}
}
\description{
Originial Subtract Autofluorescence from Median Single Colors
}
\examples{
NULL
}
