% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_tSNE.R
\name{Utility_tSNE}
\alias{Utility_tSNE}
\title{Runs tSNE (Rtsne implementation)}
\usage{
Utility_tSNE(
  x,
  sample.name,
  removestrings,
  subset,
  columns = NULL,
  notcolumns = NULL,
  subsample = NULL,
  export = FALSE,
  outpath = NULL,
  max_iter = 1000,
  perplexity = 30,
  ...
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subset}{The subset of interest from gating hierarchy}

\item{columns}{Which columns to use. Not combinable with notcolumns, use one or other.}

\item{notcolumns}{Which columns not to use. Not combinable with columns, use one or other.}

\item{subsample}{If downsampling is wanted.}

\item{outpath}{Location to store new .fcs files}

\item{max_iter}{Rtsne argument, default is set to 1000}

\item{perplexity}{Rtsne argument, default is set to 30}

\item{...}{Other Arguments to pass to Rtsne}
}
\value{
Numbers
}
\description{
Runs tSNE (Rtsne implementation)
}
\examples{
NULL
}
