% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_UMAP.R
\name{Utility_UMAP}
\alias{Utility_UMAP}
\title{Runs UMAP (uwot implementation)}
\usage{
Utility_UMAP(
  x,
  sample.name,
  removestrings,
  subset,
  columns = NULL,
  notcolumns = NULL,
  subsample = NULL,
  export = FALSE,
  outpath = NULL,
  metric = "euclidean",
  n_neighbors = 15,
  min_dist = 0.5,
  ...
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subset}{The subset of interest from gating hierarchy}

\item{columns}{Which columns to use. Not combinable with notcolumns, use one or other.}

\item{notcolumns}{Which columns not to use. Not combinable with columns,
use one or other.}

\item{subsample}{If downsampling is wanted.}

\item{export}{When set to TRUE returns fcs files to specified outpath.}

\item{outpath}{Location to store new .fcs files}

\item{metric}{umap Argument, default is set to "euclidean"}

\item{n_neighbors}{umap Argument, default is set to 15}

\item{min_dist}{umap Argument, default is set to 0.5}

\item{...}{Other arguments to pass to umap()}
}
\value{
Numbers
}
\description{
Runs UMAP (uwot implementation)
}
\examples{
NULL
}
