% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Decision.R
\name{Wetlab_Decision}
\alias{Wetlab_Decision}
\title{Takes Rest output, and referencing TheCondition csv makes decision for culture resuspensions}
\usage{
Wetlab_Decision(
  data,
  FinalConcentration_MillionperML,
  MillionCellsPerTube,
  TheConditions,
  ReturnLeftover = TRUE
)
}
\arguments{
\item{data}{The Wetlab_Rest output}

\item{FinalConcentration_MillionperML}{Desired Final Concentration in Millions}

\item{MillionCellsPerTube}{Desired number of cells in Tube}

\item{TheConditions}{A filepath or data.frame to TheCondition csv}

\item{ReturnLeftover}{Whether to return leftover cells as own line}
}
\value{
A data.frame object
}
\description{
Takes Rest output, and referencing TheCondition csv makes decision for culture resuspensions
}
\examples{
NULL
}
