% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Rest.R
\name{Wetlab_Rest}
\alias{Wetlab_Rest}
\title{Takes Concentration output, returns re-suspension amounts for Wetlab users.}
\usage{
Wetlab_Rest(
  data,
  DesiredConcentration_MillionperML,
  MaxMLperTube,
  returntype,
  outpath = NULL,
  filename = "CellResuspensions",
  ColorSelection = NULL,
  outputType = "png"
)
}
\arguments{
\item{data}{The Wetlab_Concentration output (modified correct Total mL amounts)}

\item{DesiredConcentration_MillionperML}{Final desired concentration (ex. 1, 3, etc.)}

\item{MaxMLperTube}{The Final Volume of the rest tube}

\item{returntype}{Whether to return "data", "plot" or "both"}

\item{outpath}{File path of location to store the table}

\item{filename}{File name to store the table as, ex}

\item{ColorSelection}{Default NULL, else provide desired color palette.}

\item{outputType}{Default png, alternate pdf}
}
\value{
A data.frame of resuspension measurements to get to the desired parameters
}
\description{
Takes Concentration output, returns re-suspension amounts for Wetlab users.
}
\examples{

library(flowCore)
library(flowWorkspace)
library(CytoML)
library(dplyr)
library(purrr)

File_Location <- system.file("extdata", package = "Luciernaga")
WSP_File <- list.files(File_Location, pattern=".wsp", full.names = TRUE)
ws <- open_flowjo_xml(WSP_File[1])
gs <- flowjo_to_gatingset(ws, name=1, path = File_Location)
nameKeyword <- c("GROUPNAME", "TUBENAME")

TheData <- map(.x=gs, Wetlab_Concentration, subset = "CD45+",
  nameKeyword=nameKeyword, DilutionMultiplier=100, TotalVolume=1) |>
   bind_rows()

UpdatedData <- TheData |> select(-TotalScientific, -TimeSeconds)

Results <- Wetlab_Rest(data=UpdatedData, DesiredConcentration_MillionperML=3,
 MaxMLperTube=1, returntype="data", outpath=path)

}
