% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutofluorescenceShop.R
\name{AutofluoresceShop}
\alias{AutofluoresceShop}
\title{Handling function for AF processing and tag return for the AB LabNotebook}
\usage{
AutofluoresceShop(
  x,
  visualized,
  files,
  experimentdesignation = "AB",
  template,
  GatePlots = TRUE,
  TheN = 3,
  Display = "selection",
  AFOverlap = NULL,
  ExceptFor = NULL
)
}
\arguments{
\item{x}{An experiment folder, typically being iterated through}

\item{visualized}{A vector of experiment folders already processed}

\item{files}{A list of file.path to the Unstained raw .fcs files}

\item{experimentdesignation}{Default is "AB", used to identify}

\item{template}{A file.path to the openCyto gating template for our raw
unstained .fcs files}

\item{GatePlots}{Default is TRUE, returns Utility_Gating plots for all
GatingSet files to verify gate placement worked as expected}

\item{TheN}{Selects the number of signature variants per peak detector}

\item{Display}{Default "selection" returns visual plots showing only TheN,
alternatively "all" will show all signatures before filtering in the plots}

\item{AFOverlap}{Default NULL, alternately a file.path to an AFOverlap .csv file}

\item{nameOverride}{Default FALSE, if true uses all fcs files provided regardless
if they match the experiment name string character provided in x}

\item{subsets}{Default lymphocytes, alternatively provide gate name for Luciernaga_QC to
retrieve autofluorescence signatures from.}

\item{therows}{Default 3}

\item{thecolumns}{Default 1}
}
\value{
Selected .fcs tags to the Tags folder, visualized plot
 pdf and data .csv to the Autofluorescence folder.
}
\description{
Handling function for AF processing and tag return for the AB LabNotebook
}
