% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_QC.R
\name{AveragedSignature}
\alias{AveragedSignature}
\title{Summarize a data.frame to desired stat}
\usage{
AveragedSignature(x, stats, normalize = FALSE)
}
\arguments{
\item{x}{A data.frame containing double or numeric data.}

\item{stats}{Desired Stats "mean" or "median" to pass to summarize_all}

\item{normalize}{Default FALSE, TRUE peak detector normalizes.}
}
\value{
A data.frame row of summarized data
}
\description{
Summarize a data.frame to desired stat
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
UnstainedFCSFiles <- FCS_Files[grep("Unstained", FCS_Files)]
UnstainedCells <- UnstainedFCSFiles[-grep("Beads", UnstainedFCSFiles)]
MyCytoSet <- load_cytoset_from_fcs(UnstainedCells[1],
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <-  c("DR_", "Cells", ".fcs", "-", " ")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

PopulationInterest <- gs_pop_get_data(MyGatingSet[1], subset="lymphocytes")
TheDataValues <- exprs(PopulationInterest[[1]])
TheDataValues <- data.frame(TheDataValues, check.names=FALSE)

Signature <- AveragedSignature(TheDataValues, stats="median")

}
