% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_QC.R
\name{AveragedSignature}
\alias{AveragedSignature}
\title{Summarize a data.frame to desired stat}
\usage{
AveragedSignature(x, stats)
}
\arguments{
\item{x}{A data.frame containing double or numeric data.}

\item{stats}{Desired Stats "mean" or "median" to pass to summarize_all}
}
\value{
A data.frame row of summarized data
}
\description{
Summarize a data.frame to desired stat
}
\examples{
NULL
}
