% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrightnessPlots.R
\name{BrightnessPlots}
\alias{BrightnessPlots}
\title{Visualize MFI of raw .fcs files to evaluate single color controls.}
\usage{
BrightnessPlots(thedata, input, Downsample = NULL)
}
\arguments{
\item{thedata}{A data.frame with columns Fluorophore and Detector.}

\item{input}{The location where the .fcs files are stored.}

\item{Downsample}{Reduce all clusters to just the same amount cells least represented cluster.}
}
\value{
Visualized ggplots for each fluorophore. If multiple .fcs files of same fluorophore are
 present in the same folder, it overlays them.
}
\description{
Visualize MFI of raw .fcs files to evaluate single color controls.
}
\examples{
NULL
}
