% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytosetScreen.R
\name{CytosetScreen}
\alias{CytosetScreen}
\title{Checks and filters .fcs files into compatible sublist to 
pass on to CytoSet without crashing out due to mismatch}
\usage{
CytosetScreen(files)
}
\arguments{
\item{files}{A list containing fcs files file.paths}
}
\value{
A list containing lists of compatible fcs file paths
}
\description{
Checks and filters .fcs files into compatible sublist to 
pass on to CytoSet without crashing out due to mismatch
}
