% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GateExprsIterated.R
\name{GateExprsIterated}
\alias{GateExprsIterated}
\title{Internal for RetrieveAveragedSignature, iterates on per gate
 basis to grab the corresponding averaged values for the cells
 contained within}
\usage{
GateExprsIterated(
  gs,
  sample,
  gate,
  inverse.transform = TRUE,
  excludeThese = "FSC|SSC|Time|-H"
)
}
\arguments{
\item{gs}{A gating set}

\item{inverse.transform}{Whether to inverse.transform, default
is set to TRUE}

\item{excludeThese}{Removes columns whose name contain these
 values via str_detect, default set for "FSC|SSC|Time|-H" 
which is typical for Cytek Aurora raw .fcs files}

\item{template}{The template data.frame}
}
\value{
A data.frame row containing the fluorophore name and
 the corresponding detector columns with the average signature
 value for the cells within that gate
}
\description{
Internal for RetrieveAveragedSignature, iterates on per gate
 basis to grab the corresponding averaged values for the cells
 contained within
}
\examples{
A <- 2+2

}
