% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatePlot.R
\name{GatePlot}
\alias{GatePlot}
\title{Generates called plots from Utility_GatingPlots}
\usage{
GatePlot(x, data, gtFile)
}
\arguments{
\item{x}{A specific gate, ex. "nonDebris"}

\item{data}{A GatingSet object}

\item{gtFile}{The data.table imported .csv file containing the gating template.}
}
\value{
A ggplot corresponding to the given inputs
}
\description{
Generates called plots from Utility_GatingPlots
}
\examples{
NULL
}
\keyword{internal}
