% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_DensityOverlay.R
\name{InternalDensity}
\alias{InternalDensity}
\title{Internal for Utility_DensityOverlay}
\usage{
InternalDensity(x, cs, TheFill)
}
\arguments{
\item{x}{A passed x-axis name}

\item{cs}{The passed CytoSet}

\item{TheFill}{The passed factor to fill plots by}
}
\value{
An internal value
}
\description{
Internal for Utility_DensityOverlay
}
\keyword{internal}
