% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga-package.R
\docType{package}
\name{Luciernaga-package}
\alias{Luciernaga}
\alias{Luciernaga-package}
\title{Luciernaga: Tools for the evaluation of Spectral Flow Cytometry (SFC) unmixing controls}
\description{
To learn more about using the Luciernaga package, please start by browsing the
vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidrach.github.io/Luciernaga/}
  \item Report bugs at \url{https://github.com/DavidRach/Luciernaga/issues}
}

}
\author{
\strong{Maintainer}: David Rach \email{drach@som.umaryland.edu} (\href{https://orcid.org/0000-0003-2235-5585}{ORCID})

}
\keyword{internal}
