% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga-package.R
\docType{package}
\name{Luciernaga-package}
\alias{Luciernaga}
\alias{Luciernaga-package}
\title{Luciernaga: Quality control of single color and unstained controls for Spectral Flow Cytometry}
\description{
To learn more about how to use Luciernaga, please start by browsing the vignettes:
#' @keywords internal
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavidRach/Luciernaga}
  \item Report bugs at \url{https://github.com/DavidRach/Luciernaga/issues}
}

}
\author{
\strong{Maintainer}: David Rach \email{drach@som.umaryland.edu} (\href{https://orcid.org/0000-0003-2235-5585}{ORCID})

}
