% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_BrightnessOverTime.R
\name{Luciernaga_BrigtnessOverTime}
\alias{Luciernaga_BrigtnessOverTime}
\title{Visualizes Cluster MFI over time, uses Luciernaga_QC output}
\usage{
Luciernaga_BrigtnessOverTime(data)
}
\arguments{
\item{data}{The LuciernagaQC data.frame style output with raw detector values.}
}
\value{
A ggplot2 object
}
\description{
Visualizes Cluster MFI over time, uses Luciernaga_QC output
}
\examples{
library(Luciernaga) 

}
