% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_FCSToReport.R
\name{Luciernaga_FCSToReport}
\alias{Luciernaga_FCSToReport}
\title{Visualize cosine similarity of raw .fcs files to evaluate single color
controls.}
\usage{
Luciernaga_FCSToReport(
  path,
  reference,
  stats = "median",
  LinePlots = TRUE,
  CosinePlots = TRUE,
  StackedBarPlots = TRUE,
  HeatmapPlots = TRUE,
  RetainedType,
  experiment,
  condition,
  TheSummary = TRUE
)
}
\arguments{
\item{path}{The location to the folder where the Luciernaga .fcs files are stored}

\item{reference}{A path to a .csv file or a dataframe containing Fluorophore and Detector column
information for the panel.}

\item{stats}{Whether to use the median or mean for fluorescent intensity.}

\item{LinePlots}{Return this kind of plot, default is set to TRUE}

\item{CosinePlots}{Return this kind of plot, default is set to TRUE}

\item{StackedBarPlots}{Return this kind of plot, default is set to TRUE}

\item{HeatmapPlots}{Return this kind of plot, default is set to TRUE}

\item{RetainedType}{Whether the data.frame contains "raw" or "normalized" values}

\item{experiment}{Provide directly experiment name (ex. "JAN2024")}

\item{condition}{Provide directly experiment name (ex. "JAN2024")}

\item{TheSummary}{Whether to return individual cells or summarized by stats.}
}
\value{
A data.frame compatible with LuciernagaReport()
}
\description{
Visualize cosine similarity of raw .fcs files to evaluate single color
controls.
}
\examples{
NULL
}
