% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_FolderBrightness.R
\name{Luciernaga_FolderBrightness}
\alias{Luciernaga_FolderBrightness}
\title{Sister function to FolderSignature, returns a density plot of brightness}
\usage{
Luciernaga_FolderBrightness(
  FolderPath,
  sample.name,
  StringRemoval = NULL,
  fluorophore.name,
  Verbose = FALSE,
  stats = "median",
  PanelCuts = NULL,
  normalize = TRUE
)
}
\description{
Sister function to FolderSignature, returns a density plot of brightness
}
