% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_GroupHeatmap.R
\name{Luciernaga_GroupHeatmap}
\alias{Luciernaga_GroupHeatmap}
\title{A group version of the heatmap option from Luciernaga_Plots}
\usage{
Luciernaga_GroupHeatmap(
  reports,
  nameColumn,
  cutoff = 0.01,
  returntype = "plot"
)
}
\arguments{
\item{reports}{A list of Luciernaga_QC report data objects}

\item{nameColumn}{The name of the column that differentiates between the reports}

\item{cutoff}{Proportion of cells that at least 1 report needs to exceed for retention.}

\item{returntype}{Either "plot" or underlying "data"}
}
\value{
Either a plot or underlying data
}
\description{
A group version of the heatmap option from Luciernaga_Plots
}
\examples{
NULL
}
