% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_LinearSlices.R
\name{Luciernaga_LinearSlices}
\alias{Luciernaga_LinearSlices}
\title{Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots
 the results.}
\usage{
Luciernaga_LinearSlices(
  x,
  subset,
  sample.name,
  removestrings,
  stats,
  returntype,
  probsratio = 0.1,
  output,
  desiredAF = NULL,
  legend = TRUE,
  droplowest = TRUE,
  titlename = NULL,
  returncutplot = FALSE,
  titleplot = NULL,
  inverse.transform = TRUE,
  parentSignature = NULL
)
}
\arguments{
\item{x}{A GatingSet object}

\item{subset}{The desired gating hierarchy level to look at the data}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of values to remove from the name}

\item{stats}{Whether to use "mean" or "median"}

\item{returntype}{Whether to return a "raw" or "normalized" value lineplot.}

\item{probsratio}{Ratio increments to break quantiles into, default is set
to 0.1.}

\item{output}{Whether to return "plot" or "data"}

\item{desiredAF}{Peak detector(ex. "V7-A") want to filter cells by before
slicing, argument
only used to override the main peak detector when a .fcs file has more than
 a single peak
detector, default is set to NULL}

\item{legend}{Returns the legend, default is TRUE.}

\item{droplowest}{Removes lowest bin (percentile 0), default is TRUE}

\item{titlename}{Default NULL, otherwise provide an alternate title.}

\item{returncutplot}{Default FALSE, if true returns a histogram plot with
locations
where percentile slice occured.}

\item{titleplot}{Default NULL, sets the returncutplot title}

\item{inverse.transform}{Default TRUE, whether to inverse.transform.}

\item{parentSignature}{TBD}
}
\value{
Either ggplots or the summarized data.frame object preceding
}
\description{
Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots
 the results.
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)
library(dplyr)
library(purrr)
library(stringr)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
CellSingleColorFiles <- FCS_Files[grep("Cells", FCS_Files)]
CellSingleColors <- CellSingleColorFiles[!str_detect("Unstained", CellSingleColorFiles)]

APC_Example <- CellSingleColors[grep("CD16_", CellSingleColors)]

MyCytoSet <- load_cytoset_from_fcs(APC_Example,
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <- c(".fcs")

NormalizedSlices <- Luciernaga_LinearSlices(x=MyGatingSet[1],
 subset="lymphocytes",sample.name="GUID", removestrings=removestrings,
 stats="median", returntype="normalized", probsratio=0.1, output="plot",
 desiredAF="R1-A")

}
