% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_LinearSlices.R
\name{Luciernaga_LinearSlices}
\alias{Luciernaga_LinearSlices}
\title{Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.}
\usage{
Luciernaga_LinearSlices(
  x,
  subset,
  sample.name,
  removestrings,
  stats,
  returntype,
  probsratio
)
}
\arguments{
\item{x}{A GatingSet object}

\item{subset}{The desired gating hierarchy level to look at the data}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of values to remove from the name}

\item{stats}{Whether to use "mean" or "median"}

\item{returntype}{Whether to return a "raw" or "normalized" value lineplot.}

\item{probsratio}{Ratio increments to break quantiles into, default is set to 0.1.}
}
\value{
A ggplot2 object of lineplots from the sliced .fcs file
}
\description{
Takes a Lucierna SingleColorQC .fcs, splits into percentiles, and plots the results.
}
\examples{
NULL
}
