% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_QC.R
\name{Luciernaga_QC}
\alias{Luciernaga_QC}
\title{Main Luciernaga Function, normalized on individual cell level}
\usage{
Luciernaga_QC(
  x,
  subsets,
  sample.name,
  removestrings = NULL,
  Verbose = FALSE,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  AFOverlap,
  unmixingcontroltype = "both",
  Unstained = FALSE,
  ratiopopcutoff = 0.01,
  stats = "median",
  Subtraction = "Internal",
  desiredAF = NULL,
  BeadAF = NULL,
  BeadMainAF = NULL,
  CellAF = NULL,
  CellMainAF = NULL,
  SignatureReturnNow = FALSE,
  Increments,
  LocalMaximaRatio = 0.15,
  SecondaryPeaks = 2,
  RetainedType = "normalized",
  Brightness = FALSE,
  ExportType,
  minimalfcscutoff = 0.05,
  SCData = "subtracted",
  NegativeType = "default",
  TotalNegatives = 500,
  outpath
)
}
\arguments{
\item{x}{A Gating Set object (ex. gs or gs[[1]])}

\item{subsets}{The Gating Hierarchy level you will be sampling at}

\item{sample.name}{Keyword variable which samples are stored (ex. "GUID")}

\item{removestrings}{A string of character values to remove from sample.name}

\item{Verbose}{Whether to print outputs as you go.}

\item{experiment}{Provide directly experiment name (ex. "JAN2024")}

\item{experiment.name}{Keyword variable which experiment information
is stored (ex. "TUBENAME")}

\item{condition}{Provide directly experiment name (ex. "JAN2024")}

\item{condition.name}{Keyword variable which experiment information
is stored (ex. "TUBENAME")}

\item{AFOverlap}{A data.frame or a filepath to the CSV containing the
Autofluorescence
overlap of individual fluorophores for exclusion}

\item{unmixingcontroltype}{Whether your inputs are "cells", "beads" or "both"}

\item{Unstained}{Set to True when running unstained samples that don't have
Unstained in the Name}

\item{ratiopopcutoff}{A numeric ratio for peak detector inclusion, default is
set to 0.01 all startingcells}

\item{stats}{Whether to take "mean" or "median"}

\item{Subtraction}{Whether for single color controls to use "Internal", "Internal_General" or
"External" autofluorescence}

\item{desiredAF}{Main Autofluorescence Detector (ex. "V7-A")}

\item{BeadAF}{A passed data.frame row containing the reference for bead unstained.}

\item{BeadMainAF}{The detector that corresponds to the "main" bead AF, albeit dim.}

\item{CellAF}{A passed data.frame row containing the reference for cell unstained.}

\item{CellMainAF}{The detector that corresponds to the "main" cell AF.}

\item{SignatureReturnNow}{Short circuits the function and returns signature
for specified autofluorescence.}

\item{Increments}{Rounding parameter, default is set to 0.1}

\item{LocalMaximaRatio}{Height of peaks to proceed}

\item{SecondaryPeaks}{Number of Secondary Peaks, default is set to 2.}

\item{RetainedType}{Whether to return "raw" or "normalized" values for lineplots.}

\item{Brightness}{Whether sum of detectors should be returned.}

\item{ExportType}{Whether to return "fcs", "data"}

\item{minimalfcscutoff}{A ratio determining number cells needed for .fcs export}

\item{SCData}{Whether to return "subtracted" or "raw" data for single colors}

\item{NegativeType}{Whether to append a negative pop to .fcs file, "default",
"artifical" or "samples"}

\item{TotalNegatives}{When setting NegativeType to "artificial" or "sample", how many events to add.}

\item{outpath}{Location where created .fcs and .csv files are sent}
}
\value{
Additional information to be added
}
\description{
Main Luciernaga Function, normalized on individual cell level
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)
library(dplyr)
library(purrr)
library(stringr)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
CellSingleColorFiles <- FCS_Files[grep("Cells", FCS_Files)]
CellSingleColors <- CellSingleColorFiles[!str_detect("Unstained", CellSingleColorFiles)]
MyCytoSet <- load_cytoset_from_fcs(CellSingleColors[1:2],
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <-  c("DR_", "Cells", ".fcs", "-", " ")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

FileLocation <- system.file("extdata", package = "Luciernaga")
pattern = "AutofluorescentOverlaps.csv"
AFOverlap <- list.files(path=FileLocation, pattern=pattern, full.names = TRUE)

SingleColor_Data <- map(.x=MyGatingSet[1:2], .f=Luciernaga_QC, subsets="lymphocytes",
 removestrings=removestrings, sample.name="GUID", unmixingcontroltype = "cells",
 Unstained = FALSE, ratiopopcutoff = 0.001, Verbose = FALSE, AFOverlap = AFOverlap,
 stats = "median", ExportType = "data", SignatureReturnNow = FALSE,
 outpath = TemporaryFolder, Increments=0.1, SecondaryPeaks=2,
 experiment = "FirstExperiment", condition = "ILTPanel", Subtraction = "Internal",
 CellAF=TheCellAF, SCData="subtracted", NegativeType="default") \%>\% bind_rows()

}
