% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_SignatureExternalUnstained.R
\name{Luciernaga_SignatureExternalUnstained}
\alias{Luciernaga_SignatureExternalUnstained}
\title{Wrapper for LuciernagaQC for large scale profiling. Matches unstained with single colors for respective
dates, extracts signature and plots, returns data.frame and list of patchwork objects.}
\usage{
Luciernaga_SignatureExternalUnstained(
  x,
  UnstainedList,
  FluorophoreList,
  Multiple = FALSE,
  GateTemplatePath,
  removestrings = ".fcs",
  AFOverlap
)
}
\arguments{
\item{x}{The list of Dates}

\item{UnstainedList}{The list of file.paths for the unstained files}

\item{FluorophoreList}{The list of file.paths for the fluorophore files}

\item{Multiple}{Default FALSE, if expecting multiple single color controls per unstained set TRUE}

\item{GateTemplatePath}{File.path to the openCyto gating template .csv}

\item{removestrings}{Default is ".fcs", removes from name}

\item{AFOverlap}{File.path to LuciernagaQC AFOverlap .csv to handle exceptions.}
}
\value{
A data.frame and a list of patchwork plots
}
\description{
Wrapper for LuciernagaQC for large scale profiling. Matches unstained with single colors for respective
dates, extracts signature and plots, returns data.frame and list of patchwork objects.
}
\examples{

library(Luciernaga)
}
