% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_SingleColors.R
\name{Luciernaga_SingleColors}
\alias{Luciernaga_SingleColors}
\title{Calculates the single color control matrix for a given quantile cutoffs and a given
statistic}
\usage{
Luciernaga_SingleColors(
  x,
  sample.name,
  removestrings,
  subset,
  PanelCuts,
  stats,
  SignatureView
)
}
\arguments{
\item{x}{A Gating Set Object}

\item{sample.name}{The Keyword for which the Fluorophore Name is stored}

\item{removestrings}{Values to remove from the name}

\item{subset}{A desired Gating Hierarchy level of cells to filter in}

\item{PanelCuts}{A .csv or dataframe containing columns Fluorophore, From and To
Fluorophore name should match sample.name style}

\item{stats}{Whether to use "mean" or "median"}

\item{SignatureView}{Whether to also return a normalized signature plot.}
}
\value{
A tibble row for each flow object containing the summarized data.
}
\description{
Calculates the single color control matrix for a given quantile cutoffs and a given
statistic
}
\examples{
NULL
}
