% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Tree.R
\name{Luciernaga_Tree}
\alias{Luciernaga_Tree}
\title{Select candidate Luciernaga output .fcs files for future use in unmixing.}
\usage{
Luciernaga_Tree(BrightnessFilePath, PanelPath)
}
\arguments{
\item{BrightnessFilePath}{A path to the folder the individual brightness files are in.}

\item{PanelPath}{A path to the .csv file containing the panel information. It should include only
the fluorophores captured by the BrightnessFiles}
}
\value{
A data.frame listing the candidate .fcs files for future unmixing use.
}
\description{
Select candidate Luciernaga output .fcs files for future use in unmixing.
}
\examples{
NULL
}
