% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Unmix.R
\name{Luciernaga_Unmix}
\alias{Luciernaga_Unmix}
\title{OLS unmixing of a Gating Set object}
\usage{
Luciernaga_Unmix(
  x,
  controlData,
  sample.name,
  removestrings,
  Verbose,
  addon,
  subset = "root",
  outpath,
  PanelPath,
  returnType = "fcs",
  inverse.transform = FALSE
)
}
\arguments{
\item{x}{A Gating Set object}

\item{controlData}{The matrix of single color controls generated
from Luciernaga}

\item{sample.name}{The keyword containing the fcs file name}

\item{removestrings}{A list of values to remove from name}

\item{Verbose}{For troubleshooting name after removestrings}

\item{addon}{Additional addon to append to the new .fcs file name}

\item{subset}{A gating hierarchy level to sort cells at, expression
 values retrieved
from these}

\item{outpath}{The return folder for the .fcs files}

\item{PanelPath}{Location to a panel.csv containing correct order of
fluorophores}

\item{returnType}{Whether to return "fcs" or "flowframe"}

\item{inverse.transform}{Default is FALSE, set to TRUE if data is
 already transformed
and needs to be reversed.}
}
\value{
A new .fcs file with the new columns appended
}
\description{
OLS unmixing of a Gating Set object
}
\examples{
A <- 2+2

}
