% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaReport.R
\name{LuciernagaReport}
\alias{LuciernagaReport}
\title{Converts the Luciernaga outputs into .pdf plots.}
\usage{
LuciernagaReport(
  data,
  RetainedType,
  CellPopRatio,
  outfolder,
  filename,
  LinePlots = TRUE,
  CosinePlots = TRUE,
  StackedBarPlots = TRUE,
  HeatmapPlots = TRUE,
  returntype = "patchwork"
)
}
\arguments{
\item{data}{The data.frame output from LuciernagaQC}

\item{RetainedType}{Whether the data.frame contains "raw" or "normalized" values}

\item{CellPopRatio}{What mininum ratio needed to retain cluster (ie, 0.01 = 1%, etc.)}

\item{outfolder}{The location that you want to save the .pdf output to.}

\item{filename}{The name you want to save your .pdf file as.}

\item{LinePlots}{Return this kind of plot, default is set to TRUE}

\item{CosinePlots}{Return this kind of plot, default is set to TRUE}

\item{StackedBarPlots}{Return this kind of plot, default is set to TRUE}

\item{HeatmapPlots}{Return this kind of plot, default is set to TRUE}

\item{returntype}{Passed to Utility_Patchwork for "pdf" or "patchwork" or "plots"}
}
\value{
A value to be determined later
}
\description{
Converts the Luciernaga outputs into .pdf plots.
}
\examples{
NULL
}
