% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LuciernagaTree.R
\name{LuciernagaTree}
\alias{LuciernagaTree}
\title{Select candidate Luciernaga output .fcs files for future use in unmixing.}
\usage{
LuciernagaTree(x, data)
}
\arguments{
\item{x}{A vector of the fluorophores found in data.}

\item{data}{The data.frame of Luciernaga outputs used to candidate .fcs files
for unmixing.}
}
\value{
A data.frame listing the candidate .fcs files for future unmixing use.
}
\description{
Select candidate Luciernaga output .fcs files for future use in unmixing.
}
\examples{
NULL
}
