% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameCleanup.R
\name{NameCleanUp}
\alias{NameCleanUp}
\title{Shorten a variable name}
\usage{
NameCleanUp(name, removestrings, substitutestrings)
}
\arguments{
\item{name}{The variable that you wish to shorten.}

\item{removestrings}{A list of strings to be sequentially removed from 
the name.
 An example: removestrings = c("ReferenceGroup_", ".fcs")}

\item{substitutestrings}{A data.frame containing two columns, Original
 and
Substitute.
 When a row of Original is recognized, it's replaced with substitute
 value.}
}
\value{
The shortened variable
}
\description{
Shorten a variable name
}
\examples{
name <- "DR BUV496 CD8 (Cells).fcs"
removestrings <- c("DR", "(Cells)", ".fcs", " ")
Cleaned_Name <- NameCleanUp(name, removestrings)
}
