% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameCleanup.R
\name{NameCleanUp}
\alias{NameCleanUp}
\title{Polish up single color names before unmixing}
\usage{
NameCleanUp(name, removestrings)
}
\arguments{
\item{name}{Original name}

\item{removestrings}{Strings to be removed from name (ex. c("DTR_ILT_", "cord"))}
}
\description{
Polish up single color names before unmixing
}
\examples{
NULL
}
