% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Amalgamate.R
\name{QC_Amalgamate}
\alias{QC_Amalgamate}
\title{Taking a Luciernaga_Slices (or similar data.frame) with an added Counts column, 
expands each row by the count, takes the median, and then returns either data
or a plot showing how different the average signature is from the individual ones}
\usage{
QC_Amalgamate(
  data,
  samplecolumn = "Percentiles",
  normalize = FALSE,
  countcolumn = "Count",
  returnType = "plot",
  titlename = NULL,
  linecolor = "red",
  legend = TRUE
)
}
\arguments{
\item{data}{The modified data.frame containing the sample, count and
all the normalized detector columns}

\item{samplecolumn}{The column name correspondoing to the sample column,
default is Percentiles}

\item{countcolumn}{The column name corresponding to the count column,
default Count}

\item{returnType}{Whether to return plot (default) or data}

\item{titlename}{Default is NULL, specify a title if desired}

\item{linecolor}{Default is red, specify alternate color for the Average signature}

\item{legend}{Whether to return legend, default is TRUE}
}
\value{
Either a plot or the underlying data
}
\description{
Taking a Luciernaga_Slices (or similar data.frame) with an added Counts column, 
expands each row by the count, takes the median, and then returns either data
or a plot showing how different the average signature is from the individual ones
}
\examples{

A <- 2+2

}
