% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_Conversions_Chorus.R
\name{QC_ChorusPDF}
\alias{QC_ChorusPDF}
\title{Processes BD Chorus QC PDF files into .csv files}
\usage{
QC_ChorusPDF(x, returnPreference = "Imaging")
}
\arguments{
\item{x}{A file.path to the desired QC pdf}
}
\description{
Processes BD Chorus QC PDF files into .csv files
}
\examples{
A <- 2+2

}
