% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_Conversions_FlowSight.R
\name{QC_FlowSightPDF}
\alias{QC_FlowSightPDF}
\title{Parses the two-page Amnis FlowSight QC report, returning tidy dataframe.}
\usage{
QC_FlowSightPDF(x)
}
\arguments{
\item{x}{A file.path to the respective QC report pdf}
}
\description{
Parses the two-page Amnis FlowSight QC report, returning tidy dataframe.
}
\examples{
A <- 2+2
}
