% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_LibraryParse.R
\name{QC_LibraryParse}
\alias{QC_LibraryParse}
\title{Parses Library Reference Control .XML files and returns}
\usage{
QC_LibraryParse(
  x,
  returntype,
  references = TRUE,
  myfactor = "Fluorophore",
  namefactor = "Sample"
)
}
\arguments{
\item{x}{An .XML file}

\item{returntype}{What to return "data" or "plots"}

\item{references}{Plot argument, adds red reference signature}

\item{myfactor}{Plot argument, data column to group by for plotting. Default "Fluorophore".}

\item{namefactor}{Plot argument, data column name added to Plot Title.}
}
\value{
TBD
}
\description{
Parses Library Reference Control .XML files and returns
}
\examples{
Folder_Location <- system.file("extdata", package = "Luciernaga")
XML_Pattern <- ".XML$"
XML_Files <- list.files(path = Folder_Location, pattern = XML_Pattern,
                        full.names = TRUE, recursive = FALSE)
SinglePlot <- QC_LibraryParse(XML_Files[2], returntype="plots", references=FALSE)
}
