% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_LibraryParse.R
\name{QC_LibraryParse}
\alias{QC_LibraryParse}
\title{Parses Library Reference Control .XML files and returns}
\usage{
QC_LibraryParse(x, returntype, references = TRUE)
}
\arguments{
\item{x}{An .XML file}

\item{returntype}{What to Return}

\item{references}{Whether to plot reference files}
}
\value{
TBD
}
\description{
Parses Library Reference Control .XML files and returns
}
\examples{
NULL

}
