% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_LibraryParse.R
\name{QC_LibraryParse}
\alias{QC_LibraryParse}
\title{Parses Library Reference Control .XML files and returns}
\usage{
QC_LibraryParse(x, returntype, references = TRUE)
}
\arguments{
\item{x}{An .XML file}

\item{returntype}{What to return "dataframe" or "plots"}

\item{references}{Whether to add reference fluorophore signature in red}
}
\value{
TBD
}
\description{
Parses Library Reference Control .XML files and returns
}
\examples{

Folder_Location <- system.file("extdata", package = "Luciernaga")
XML_Pattern <- ".XML$"
XML_Files <- list.files(path = Folder_Location, pattern = XML_Pattern,
                        full.names = TRUE, recursive = FALSE)
SinglePlot <- QC_LibraryParse(XML_Files[2], returntype="plots", references=FALSE)
}
