% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Retrieval.R
\name{QC_Retrieval}
\alias{QC_Retrieval}
\title{Returns the Detector and Laser settings from individual CytoSet files}
\usage{
QC_Retrieval(x, sample.name)
}
\arguments{
\item{x}{A CytoSet object when mapped, or its frame (cs[1])}

\item{sample.name}{The keyword value that distinguishes individual .fcs files}
}
\value{
A dataframe row
}
\description{
Returns the Detector and Laser settings from individual CytoSet files
}
\examples{
NULL
}
