% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Retrieval.R
\name{QC_Retrieval}
\alias{QC_Retrieval}
\title{Returns Detector Gains, Laser Delay and Scalings from individual .fcs files in a
CytoSet.}
\usage{
QC_Retrieval(x, sample.name)
}
\arguments{
\item{x}{A CytoSet (when mapped) or an individual cytoset, example (x=MyCytoSet[[1]])}

\item{sample.name}{The keyword value that distinguishes individual .fcs files}
}
\value{
A dataframe row
}
\description{
Returns Detector Gains, Laser Delay and Scalings from individual .fcs files in a
CytoSet.
}
\examples{
library(flowWorkspace)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Pattern <- ".fcs$"
FCS_Files <- list.files(path = File_Location, pattern = FCS_Pattern,
full.names = TRUE, recursive = FALSE)
QCBeads <- FCS_Files[grep("Before", FCS_Files)]
MyCytoSet <- load_cytoset_from_fcs(QCBeads[1], truncate_max_range = FALSE,
transformatio = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)

SingleSpecimen <- QC_Retrieval(x=MyGatingSet[[1]], sample.name="TUBENAME")

}
