% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Screen.R
\name{QC_Screen}
\alias{QC_Screen}
\title{Takes Luciernaga_QC data output, returns amalgamate plots, used
to screen QC beads for major issues that notably shift the average}
\usage{
QC_Screen(
  data,
  groupColumns,
  clusterColumn = "Cluster",
  filterDate = NULL,
  cutoff = 0.005,
  normalize = FALSE,
  returnType = "pdf",
  outpath,
  filename = "BeadQCVisualized",
  therows = 3,
  thecolumns = 3
)
}
\arguments{
\item{data}{The LuciernagaQC data output}

\item{groupColumns}{Names of columns you want to combine to make the 
name being used for comparison, ex: c("Sample", "Condition")}

\item{clusterColumn}{Default is "Cluster"}

\item{filterDate}{Used to quickly return a screened plot, use the plot title
name in it's entirety and set returnType to "plot"}

\item{cutoff}{Default is 0.005}

\item{normalize}{Default is FALSE}

\item{returnType}{Default is pdf, alternatively plot}

\item{outpath}{The desired storage folder}

\item{filename}{The desired filename for the output to be saved as}

\item{therows}{For pdf output, number of rows per page, default 3}

\item{thecolumns}{For pdf output, number of columns per page, default 3}
}
\value{
Either individual plots or a pdf desired location
}
\description{
Takes Luciernaga_QC data output, returns amalgamate plots, used
to screen QC beads for major issues that notably shift the average
}
