% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_SimilarFluorophores.R
\name{QC_SimilarFluorophores}
\alias{QC_SimilarFluorophores}
\title{Queries fluorophore and returns similar fluorophores.}
\usage{
QC_SimilarFluorophores(
  TheFluorophore,
  NumberDetectors,
  NumberHits,
  returnPlots = FALSE
)
}
\arguments{
\item{NumberHits}{Number of most similar fluorophores by cosine.}

\item{x}{Name of Fluorophore You Want to Examine}

\item{data}{The dataframe object from QC_LibraryParse}
}
\value{
A dataframe of similar fluorophores
}
\description{
Queries fluorophore and returns similar fluorophores.
}
\examples{
NULL
}
