% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_UserLibraries.R
\name{QC_UserLibraries}
\alias{QC_UserLibraries}
\title{Takes data.frame of QC_LibraryParse, and iterating over individual users
returns plots}
\usage{
QC_UserLibraries(
  x,
  Data,
  NameAppend,
  outpath,
  references = TRUE,
  thecolumns = 3,
  therows = 4,
  width = 7,
  height = 9,
  saveCSV = TRUE
)
}
\arguments{
\item{x}{Mapped individual name present within the Creator Column}

\item{Data}{The data.frame object working from}

\item{NameAppend}{Name to add to end of the file}

\item{outpath}{Desired storage location}

\item{references}{Whether to include red reference signatures}

\item{thecolumns}{Passed to Patchwork, desired number of columns}

\item{therows}{Passed to Patchwork, desired number of rows}

\item{width}{Passed to Patchwork, page width}

\item{height}{Passed to Patchwork, page height}

\item{saveCSV}{Whether to return a .csv of underlying data, default is TRUE}
}
\value{
A pdf of the plots and maybe a csv
}
\description{
Takes data.frame of QC_LibraryParse, and iterating over individual users
returns plots
}
\examples{
NULL
}
