% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_WhatsThis.R
\name{QC_WhatsThis}
\alias{QC_WhatsThis}
\title{Querries reference signatures and returns most similar fluorophores by cosine similarity.}
\usage{
QC_WhatsThis(x, data, NumberHits, returnPlots)
}
\arguments{
\item{x}{Name of Fluorophore/Sample you want to filter for}

\item{data}{A data.frame object from QC_LibraryParse containing Fluorophore name column and numeric detector columns.}

\item{NumberHits}{Number of most similar fluorophores by cosine.}
}
\value{
A dataframe of similar fluorophores
}
\description{
Querries reference signatures and returns most similar fluorophores by cosine similarity.
}
\examples{
NULL
}
