% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_WhatsThis.R
\name{QC_WhatsThis}
\alias{QC_WhatsThis}
\title{Querries reference signatures and returns most similar fluorophores by cosine similarity.}
\usage{
QC_WhatsThis(x, data, NumberHits, returnPlots = FALSE)
}
\arguments{
\item{x}{Name in the Sample column you want to filter for}

\item{data}{A data.frame object from QC_LibraryParse containing Fluorophore name column and numeric detector columns.}

\item{NumberHits}{Number of most similar fluorophores by cosine.}

\item{returnPlots}{Whether to return signature plots, default is set to FALSE.}
}
\value{
A dataframe of similar fluorophores
}
\description{
Querries reference signatures and returns most similar fluorophores by cosine similarity.
}
\examples{
library(dplyr)
Folder_Location <- system.file("extdata", package = "Luciernaga")
XML_Pattern <- ".XML$"
XML_Files <- list.files(path = Folder_Location, pattern = XML_Pattern,
                        full.names = TRUE, recursive = FALSE)
LibraryData <- QC_LibraryParse(XML_Files[2], returntype="dataframe", references=FALSE)
Data <- LibraryData \%>\% select(-Sample, -Creator, -Date) \%>\%
 rename(Sample=Fluorochrome)
TheFluorophore <- Data \%>\% pull(Sample)

Results <- QC_WhatsThis(x=TheFluorophore, data=Data, NumberHits = 10, returnPlots=FALSE)
}
