% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCVfromFCS.R
\name{RCVfromFCS}
\alias{RCVfromFCS}
\title{Calculate the RCV value for each marker for a specified GatingSet subset}
\usage{
RCVfromFCS(x, subset, sample.name)
}
\arguments{
\item{x}{The Gating Set object being iterated}

\item{subset}{The Gating Set subset for population of interest}

\item{sample.name}{The keyword corresponding to the samples name}
}
\value{
A data.frame row with the marker RCVs as a proportion.
}
\description{
Calculate the RCV value for each marker for a specified GatingSet subset
}
