% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Retrieval.R
\name{RetrievalMerge}
\alias{RetrievalMerge}
\title{Internal for QC_Retrieval}
\usage{
RetrievalMerge(x, y, TheData)
}
\arguments{
\item{x}{Passed argument 1}

\item{y}{Passed argument 2}

\item{TheData}{The datset}
}
\description{
Internal for QC_Retrieval
}
\keyword{internal}
