% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAverageSignature.R
\name{RetrieveAverageSignature}
\alias{RetrieveAverageSignature}
\title{Helper function, retrieves the average signature for each gate specified in
the template}
\usage{
RetrieveAverageSignature(
  template,
  gs,
  inverse.transform = TRUE,
  excludeThese = "FSC|SSC|Time|-H"
)
}
\arguments{
\item{template}{The template data.frame}

\item{gs}{A gating set}

\item{inverse.transform}{Whether to inverse.transform, default is set to TRUE}

\item{excludeThese}{Removes columns whose name contain these values via str_detect, 
default set for "FSC|SSC|Time|-H" which is typical for Cytek Aurora raw .fcs files}
}
\value{
An updated template data.frame containing the average for each detector
}
\description{
Helper function, retrieves the average signature for each gate specified in
the template
}
\examples{
 A <- 2+2

}
