% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexReduction.R
\name{SingleUnmix}
\alias{SingleUnmix}
\title{Carries out the initial unmixing using just the single-color controls own reference signature.}
\usage{
SingleUnmix(
  matrix,
  matrix_columnname = "Fluorophore",
  gs,
  subset = "root",
  outpath,
  inverse.transform = TRUE,
  sample.name = "TUBENAME"
)
}
\arguments{
\item{matrix}{The signature matrix for the respective single-color unmixing controls}

\item{matrix_columnname}{The column name used to match respective metadata in the GatingSet. 
Default is Fluorophore.}

\item{gs}{The GatingSet containing the raw single-color unmixing controls}

\item{subset}{The gate with the events you want to unmix. Default is root.}

\item{outpath}{The location to store the unmixed .fcs files.}

\item{inverse.transform}{Default equals TRUE}

\item{sample.name}{Keyword(s) to grab from the cytoset for renaming}
}
\value{
Unmixed single-color unmixing controls for use in stain reduction calculations.
}
\description{
Carries out the initial unmixing using just the single-color controls own reference signature.
}
\examples{
A <- 2 + 2

}
