% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Luciernaga_Lists.R
\name{Subplots}
\alias{Subplots}
\title{Internal for Luciernaga_Lists}
\usage{
Subplots(i, data)
}
\arguments{
\item{i}{Passed Indicies}

\item{data}{The Transposed List of Lists}
}
\description{
Internal for Luciernaga_Lists
}
\keyword{internal}
