% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationCheck.R
\name{TransformationCheck}
\alias{TransformationCheck}
\title{Check the transformation settings to visualize .fcs files}
\usage{
TransformationCheck(
  x,
  thesecolumns = NULL,
  TransformationChoice = "flowjo_biexp",
  channelRange = 256,
  maxValue = 1e+06,
  pos = 4.5,
  neg = 1,
  widthBasis = -500,
  Multiple = FALSE,
  returnType,
  optionalY = "BUV805-A",
  optionalX = "BUV496-A",
  optionalOutpath = NULL,
  optionalBins = 100
)
}
\arguments{
\item{x}{Iterated in GatingSet object}

\item{thesecolumns}{Choice of channels to transform}

\item{TransformationChoice}{Default flowjo_biexp_trans}

\item{channelRange}{Argument}

\item{maxValue}{Argument}

\item{pos}{Argument}

\item{neg}{Argument}

\item{widthBasis}{Argument}

\item{Multiple}{Argument}

\item{returnType}{Argument}
}
\value{
A visualization to check transformation setting
}
\description{
Check the transformation settings to visualize .fcs files
}
\examples{
path <- system.file("extdata", package = "Luciernaga")
files <- list.files(path, pattern=".fcs", full.names=TRUE)

}
