% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ColAppend.R
\name{Utility_ColAppend}
\alias{Utility_ColAppend}
\title{Add Dimensionality Visualized parameters to raw .fcs files}
\usage{
Utility_ColAppend(ff, DF, columnframe, shift = FALSE)
}
\arguments{
\item{ff}{A realized_view object from flowWorkspace}

\item{DF}{The maybe downsampled exprs}

\item{columnframe}{The dimensionality visualized data.frame object to be added}

\item{shift}{Whether to shift values to non-zero, default is set to FALSE,}
}
\value{
A new flow_frame object.
}
\description{
Add Dimensionality Visualized parameters to raw .fcs files
}
\examples{

library(BiocGenerics)
library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)
library(dplyr)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
Unmixed_FullStained <- FCS_Files[grep("Unmixed", FCS_Files)]
UnmixedFCSFiles <- Unmixed_FullStained[1]
UnmixedCytoSet <- load_cytoset_from_fcs(UnmixedFCSFiles[1],
  truncate_max_range = FALSE,transformation = FALSE)
UnmixedGatingSet <- GatingSet(UnmixedCytoSet)
Markers <- colnames(UnmixedCytoSet)
KeptMarkers <- Markers[-grep("Time|FS|SC|SS|Original|-W$|-H$|AF", Markers)]
MyBiexponentialTransform <- flowjo_biexp_trans(channelRange = 256,
  maxValue = 1000000,pos = 4.5, neg = 0, widthBasis = -1000)
TransformList <- transformerList(KeptMarkers, MyBiexponentialTransform)
UnmixedGatingSet <- flowWorkspace::transform(UnmixedGatingSet, TransformList)
FileLocation <- system.file("extdata", package = "Luciernaga")
UnmixedGates <- fread(file.path(path = FileLocation, pattern = 'GatesUnmixed.csv'))
UnmixedGating <- gatingTemplate(UnmixedGates)
gt_gating(UnmixedGating, UnmixedGatingSet)

removestrings <-  c("DTR_", ".fcs")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

ff <- gs_pop_get_data(UnmixedGatingSet[1], subsets="live", inverse.transform = FALSE)
BeforeParameters <- ff[[1, returnType = "flowFrame"]]
MainDataFrame <- as.data.frame(exprs(ff[[1]]), check.names = FALSE)
NewData <- MainDataFrame \%>\% mutate(ExposureStatus = sample(1:3, n(), replace = TRUE))
NewData <- NewData \%>\% select(ExposureStatus)
AfterParameters <- Utility_ColAppend(ff=ff, DF=MainDataFrame, columnframe = NewData)

}
