% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ColAppend.R
\name{Utility_ColAppend}
\alias{Utility_ColAppend}
\title{Add Dimensionality Visualized parameters to raw .fcs files}
\usage{
Utility_ColAppend(ff, DF, columnframe, shift = FALSE)
}
\arguments{
\item{ff}{A realized_view object from flowWorkspace}

\item{DF}{The maybe downsampled exprs}

\item{columnframe}{The dimensionality visualized data.frame object to be added}

\item{shift}{Whether to shift values to non-zero, default is set to FALSE,}
}
\value{
A new flow_frame object.
}
\description{
Add Dimensionality Visualized parameters to raw .fcs files
}
\examples{
NULL
}
