% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Concatinate.R
\name{Utility_Concatinate}
\alias{Utility_Concatinate}
\title{Concatinate a gs based on subset and subsample}
\usage{
Utility_Concatinate(
  gs,
  sample.name,
  removestrings,
  subsets,
  subsample,
  newName,
  outpath
)
}
\arguments{
\item{gs}{A gating set object}

\item{sample.name}{Keyword specifying sample name}

\item{removestrings}{Value to be removed from sample name}

\item{subsets}{The gating hierarchy subset you want to include}

\item{subsample}{Total number of events to subsample from each specimen}

\item{newName}{File Name for the Concatenate File}

\item{outpath}{Location to store the concatenated file}
}
\value{
A concatenated .fcs file to new location
}
\description{
Concatinate a gs based on subset and subsample
}
\examples{
Not applicable at this time
}
