% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Concatinate.R
\name{Utility_Concatinate}
\alias{Utility_Concatinate}
\title{Concatenate a gs based on subset and optional subsample}
\usage{
Utility_Concatinate(
  gs,
  sample.name,
  removestrings,
  subsets,
  subsample = NULL,
  ReturnType,
  newName,
  outpath = NULL,
  export = FALSE,
  inverse.transform = FALSE
)
}
\arguments{
\item{gs}{A gating set object}

\item{sample.name}{Keyword specifying sample name}

\item{removestrings}{Value to be removed from sample name}

\item{subsets}{The gating hierarchy subset you want to include}

\item{subsample}{Default is set to NULL. If down-sample is desired, Total number of events to subsample from each specimen}

\item{ReturnType}{Whether to return as a "data.frame", a "flow.frame", or export to an "fcs" file to the outpath location}

\item{newName}{File Name for the Concatenate File}

\item{outpath}{Location to store the concatenated file}

\item{export}{Whether to export as a .fcs file.}

\item{inverse.transform}{Whether to reverse the GatingSet Transform on the data, default is set to FALSE.}
}
\value{
A concatenated .fcs file to new location
}
\description{
Concatenate a gs based on subset and optional subsample
}
\examples{
NULL
}
