% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Concatinate.R
\name{Utility_Concatinate}
\alias{Utility_Concatinate}
\title{Concatenate a gs based on subset and optional subsample}
\usage{
Utility_Concatinate(
  gs,
  sample.name,
  removestrings,
  subsets,
  subsample = NULL,
  ReturnType,
  newName,
  outpath = NULL,
  export = FALSE,
  inverse.transform,
  metadataCols = NULL,
  Override = FALSE,
  DataOverride = NULL
)
}
\arguments{
\item{gs}{A gating set object}

\item{sample.name}{Keyword specifying sample name}

\item{removestrings}{Value to be removed from sample name}

\item{subsets}{The gating hierarchy subset you want to include}

\item{subsample}{Default is set to NULL. If down-sample is desired, Total number of
events to subsample from each specimen}

\item{ReturnType}{Whether to return as a "data.frame", a "flow.frame", or export
to an "fcs" file to the outpath location}

\item{newName}{File Name for the Concatenate File}

\item{outpath}{Location to store the concatenated file}

\item{export}{Whether to export as a .fcs file.}

\item{inverse.transform}{Whether to reverse the GatingSet Transform on the data,
default is set to FALSE.}

\item{metadataCols}{column names from pData to append as metadata for the .fcs, default NULL}
}
\value{
A concatenated data.frame, flow.frame or fcs file, with reference .csv for
the specimen information
}
\description{
Concatenate a gs based on subset and optional subsample
}
\examples{

library(BiocGenerics)
library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
Unmixed_FullStained <- FCS_Files[grep("Unmixed", FCS_Files)]
UnmixedFCSFiles <- Unmixed_FullStained[1:2]
UnmixedCytoSet <- load_cytoset_from_fcs(UnmixedFCSFiles[1:2],
  truncate_max_range = FALSE,transformation = FALSE)
UnmixedGatingSet <- GatingSet(UnmixedCytoSet)
Markers <- colnames(UnmixedCytoSet)
KeptMarkers <- Markers[-grep("Time|FS|SC|SS|Original|-W$|-H$|AF", Markers)]
MyBiexponentialTransform <- flowjo_biexp_trans(channelRange = 256,
  maxValue = 1000000,pos = 4.5, neg = 0, widthBasis = -1000)
TransformList <- transformerList(KeptMarkers, MyBiexponentialTransform)
UnmixedGatingSet <- flowWorkspace::transform(UnmixedGatingSet, TransformList)
FileLocation <- system.file("extdata", package = "Luciernaga")
UnmixedGates <- fread(file.path(path = FileLocation, pattern = 'GatesUnmixed.csv'))
UnmixedGating <- gatingTemplate(UnmixedGates)
gt_gating(UnmixedGating, UnmixedGatingSet)

removestrings <-  c("DTR_", ".fcs")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

ConcatinatedReturn <- Utility_Concatinate(gs=UnmixedGatingSet,
  sample.name = "GROUPNAME", removestrings=removestrings,
  subsets="live", subsample = 2000, ReturnType = "flow.frame",
  newName = "MyConcatinatedFile", outpath = StorageLocation,
   export = FALSE, inverse.transform=TRUE)

}
