% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_DensityOverlay.R
\name{Utility_DensityOverlay}
\alias{Utility_DensityOverlay}
\title{Generates an overlay plot containing all objects in the GatingSet}
\usage{
Utility_DensityOverlay(
  gs,
  subset,
  inverse.transform = FALSE,
  TheX = NULL,
  TheFill,
  returntype,
  outpath,
  filename,
  therows = 3,
  thecolumns = 3,
  width = 7,
  height = 9
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{subset}{The desired Gating Hierarchy node (ex. "lymphocytes")}

\item{inverse.transform}{Default is FALSE, reverses transformation applied to
GatingSet as its converted to a CytoSet}

\item{TheX}{A desired marker to plot, leave NULL for all markers}

\item{TheFill}{A desired marker to color individual specimens by (as named in pData)}

\item{returntype}{Whether to return a "pdf", "patchwork" or "plots"}

\item{outpath}{Desired storage location}

\item{filename}{The file name for the new .pdf}

\item{therows}{The desired number of rows per page}

\item{thecolumns}{The desired number of columns per page}

\item{width}{The desired page width}

\item{height}{The desired page height}
}
\value{
A pdf file or ggplot object
}
\description{
Generates an overlay plot containing all objects in the GatingSet
}
\examples{
NULL
}
