% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Downsample.R
\name{Utility_Downsample}
\alias{Utility_Downsample}
\title{Downsample a gs based on subset}
\usage{
Utility_Downsample(x, sample.name, removestrings, subsets, subsample)
}
\arguments{
\item{sample.name}{Keyword specifying sample name}

\item{removestrings}{Value to be removed from sample name}

\item{subsets}{The gating hierarchy subset you want to include}

\item{subsample}{Total number of events to subsample from each specimen}

\item{gs}{A gating set object}

\item{newName}{File Name for the Concatenate File}

\item{outpath}{Location to store the concatenated file}
}
\value{
A concatenated .fcs file to new location
}
\description{
Downsample a gs based on subset
}
\examples{
NULL
}
