% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_GatingPlots.R
\name{Utility_GatingPlots}
\alias{Utility_GatingPlots}
\title{Check gate placement for individual .fcs files in a GatingSet}
\usage{
Utility_GatingPlots(
  x,
  sample.name,
  removestrings,
  subset = "root",
  gtFile = NULL,
  DesiredGates = NULL,
  outpath = NULL,
  returnType,
  bins = 270,
  therows = 2,
  thecolumns = 2,
  width = 7,
  height = 9,
  clearance = 0.2,
  optionalX = NULL,
  optionalY = NULL,
  optionalGate = NULL,
  ...
)
}
\arguments{
\item{x}{A GatingSet object}

\item{sample.name}{The .fcs keyword that contains an unique name for that sample}

\item{removestrings}{A string of character values to remove from the sample.name}

\item{subset}{The GatingSet subset that you want to visualize for data plotting,
"root" is the default.}

\item{gtFile}{The data.table imported .csv file containing the gating template.}

\item{DesiredGates}{A vector of gates that you want plotted, for example
Desired <- c("nonDebris, "lymphocytes")}

\item{outpath}{Location to store the generated .pdf file}

\item{returnType}{Whether to return "pdf", "patchwork" or "plots".}

\item{bins}{Argument to geom_hex for number of bins to visualize the plotted
data density.}

\item{therows}{Number of desired rows for the .pdf file}

\item{thecolumns}{Number of desired columns for the .pdf file}

\item{width}{Desired page width}

\item{height}{Desired page height}

\item{clearance}{A buffer area around the plot edge}

\item{optionalX}{When gtFile is NULL, provides x-axis argument for the subset gated population}

\item{optionalY}{When gtFile is NULL, provides y-axis argument for the subset gated population}

\item{...}{Additional Arguments to NameForSample to derive filename.}
}
\value{
Additional information to be added
}
\description{
Check gate placement for individual .fcs files in a GatingSet
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
UnstainedFCSFiles <- FCS_Files[grep("Unstained", FCS_Files)]
UnstainedCells <- UnstainedFCSFiles[-grep("Beads", UnstainedFCSFiles)]
MyCytoSet <- load_cytoset_from_fcs(UnstainedCells[1],
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <-  c("DR_", "Cells", ".fcs", "-", " ")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

IndividualPlot <- Utility_GatingPlots(x=MyGatingSet[[1]],
  sample.name = "GUID",removestrings = removestrings,
  gtFile = MyGates, DesiredGates = NULL,
  outpath = StorageLocation, returnType = "patchwork")

}
