% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_GatingPlots.R
\name{Utility_GatingPlots}
\alias{Utility_GatingPlots}
\title{Check gate placement for individual .fcs files in a GatingSet}
\usage{
Utility_GatingPlots(
  x,
  sample.name,
  removestrings,
  subsets = "root",
  gtFile,
  DesiredGates = NULL,
  thecolumn = 2,
  therows = 2,
  bins,
  clearance,
  outpath = NULL,
  export = NULL,
  ...
)
}
\arguments{
\item{x}{A GatingSet object}

\item{sample.name}{The .fcs keyword that contains an unique name for that sample}

\item{removestrings}{A string of character values to remove from the sample.name}

\item{subsets}{The GatingSet subset that you want to visualize for data plotting, "root" is the default.}

\item{gtFile}{The data.table imported .csv file containing the gating template.}

\item{DesiredGates}{A vector of gates that you want plotted, for example Desired <- c("nonDebris, "lymphocytes")}

\item{therows}{Number of desired rows for the .pdf file}

\item{bins}{Argument to geom_hex for number of bins to visualize the plotted data density.}

\item{clearance}{Area around edge}

\item{outpath}{Location to store the generated .pdf file}

\item{export}{Whether to return as a .pdf file or as individual patchwork ggplot objects.}

\item{experiment.name}{The .fcs keyword that contains the name of the experiment}

\item{experiment}{Directly provide the name of the experiment (alternative to experiment.name)}

\item{condition.name}{The .fcs keyword that contains the name of the condition.}

\item{condition}{Directly provide the name of the condition (alternative to condition.name)}

\item{thecolumns}{Number of desired columns for the .pdf file}
}
\value{
Additional information to be added
}
\description{
Check gate placement for individual .fcs files in a GatingSet
}
\examples{
NULL
}
