% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_GatingPlots.R
\name{Utility_GatingPlots}
\alias{Utility_GatingPlots}
\title{Check gate placement for individual .fcs files}
\usage{
Utility_GatingPlots(
  x,
  sample.name,
  removestrings,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  subsets,
  gtFile,
  column,
  bins,
  clearance,
  outpath,
  export = NULL
)
}
\arguments{
\item{x}{A Gating Set object}

\item{sample.name}{The .fcs keyword containing the samples name}

\item{removestrings}{The values you want removed from sample.name}

\item{experiment}{Provide an alternative experiment name}

\item{experiment.name}{The .fcs keyword containing the experiment value}

\item{condition}{Provide an alternative condition name}

\item{condition.name}{The .fcs keyword containing the condition name}

\item{subsets}{The subset level you want the data plotted at. "root" as default}

\item{gtFile}{The imported data.table containing the gate information.}

\item{column}{Residual superceeded}

\item{bins}{Number of geom hexbins to bin the plotted data for visualization.}

\item{clearance}{Area around edge}

\item{outpath}{The location to send the generated .pdf file}

\item{export}{Whether to export as a .pdf file to the outpath, TRUE, FALSE}
}
\value{
Additional information to be added
}
\description{
Check gate placement for individual .fcs files
}
\examples{
NULL
}
