% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_GeneralGating.R
\name{Utility_GeneralGating}
\alias{Utility_GeneralGating}
\title{An internal function that generates ggplots for given channels}
\usage{
Utility_GeneralGating(
  x,
  name,
  ff,
  yValue,
  clearance,
  bins,
  columnlist,
  TheDF,
  gatelines,
  reference = NULL
)
}
\arguments{
\item{x}{Passed channel}

\item{name}{The name that matches to the data.frame and gs object,
ex. altered.name}

\item{ff}{The dataframe data for that sample}

\item{yValue}{What wanted on the yaxis}

\item{clearance}{Space around the plot}

\item{bins}{Bins for which the plotted cells will be divided into
providing granularity}

\item{columnlist}{list all channels with x removed}

\item{TheDF}{the external limits settings}

\item{gatelines}{whether to plot the ModernCutoffLines from reference
dataframe}

\item{reference}{location ModernCutoff dataframe.}
}
\value{
A value to be determined later
}
\description{
An internal function that generates ggplots for given channels
}
\examples{
NULL
}
\keyword{internal}
