% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_IterativeGating.R
\name{Utility_IterativeGating}
\alias{Utility_IterativeGating}
\title{Take a Gating Set, and return individual plots for specified node}
\usage{
Utility_IterativeGating(
  x,
  subset,
  gate,
  xValue,
  yValue,
  sample.name,
  remove.strings,
  bins
)
}
\arguments{
\item{x}{A Gating Set Object}

\item{subset}{Desired node of cells to plot}

\item{gate}{Desired gate to show on the same plot as the subset, else NULL}

\item{xValue}{Desired x axis}

\item{yValue}{Desired y axis}

\item{sample.name}{Keyword under which the sample name is stored}

\item{remove.strings}{Character values to remove from the name}

\item{bins}{Geo_hex bins for the dots}
}
\value{
A list of ggplot objects
}
\description{
Take a Gating Set, and return individual plots for specified node
}
\examples{
NULL
}
