% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_NbyNPlots.R
\name{Utility_NbyNPlots}
\alias{Utility_NbyNPlots}
\title{Generate NxN plots for one or all fluorophores.}
\usage{
Utility_NbyNPlots(
  x,
  sample.name,
  removestrings,
  experiment = NULL,
  experiment.name = NULL,
  condition = NULL,
  condition.name = NULL,
  marginsubset,
  gatesubset,
  ycolumn,
  bins,
  clearance,
  gatelines,
  reference = NULL,
  outpath,
  returntype,
  width = 9,
  height = 7
)
}
\arguments{
\item{x}{A GatingSet object (ex. gs or gs[[1]])}

\item{sample.name}{Keyword for which samples names are stored (ex. "GUID")}

\item{removestrings}{A string of characters to remove from name (
ex. c("DR_ILT_2023_", "Cord"))}

\item{experiment}{Provide directly an experiment name (ex. "Jan2024")}

\item{experiment.name}{Keyword for which experiment name is stored (
ex. "GROUPNAME")}

\item{condition}{Provide a condition name}

\item{condition.name}{The keyword in the .fcs file storing the condition.name}

\item{marginsubset}{The Gating Hierarchy level from which values will be used to
estimate the plot margins (ex. "nonDebris")}

\item{gatesubset}{The Gating Hierarchy level of the cells that you want to see
plotted (ex. "lymph")}

\item{ycolumn}{The ycolumn that you want to see everything plotted by (ex. "APC-A")
or ALL to see all comparisons}

\item{bins}{Bins for which the plotted cells will be divided into providing
granularity}

\item{clearance}{The additional ratio added to the margins to avoid clipping
main population but exclude outliers.}

\item{gatelines}{Whether to add estimated gate cutoff lines}

\item{reference}{Reference for the gate cutoff lines}

\item{outpath}{Location for which to store the generated .pdf}

\item{returntype}{Whether to return as "pdf", "patchwork" or "plots"}

\item{width}{Desired page width for a pdf, default is 9 inches.}

\item{height}{Desired page height for a pdf, default is 7 inches}
}
\value{
A value to be determined later
}
\description{
Generate NxN plots for one or all fluorophores.
}
\examples{
NULL
}
