% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_PaCMAP.R
\name{Utility_PaCMAP}
\alias{Utility_PaCMAP}
\title{Runs PaCMAP via Basilisk}
\usage{
Utility_PaCMAP(x, sample.name, removestrings, subsets, columns, subsample)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subsets}{The subset of interest from gating hierarchy}

\item{columns}{A subset of columns to pass instead}

\item{subsample}{If downsampling is wanted.}
}
\value{
Numbers
}
\description{
Runs PaCMAP via Basilisk
}
\examples{
Not applicable
}
