% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_PaCMAP.R
\name{Utility_PaCMAP}
\alias{Utility_PaCMAP}
\title{Runs PaCMAP via Basilisk}
\usage{
Utility_PaCMAP(
  x,
  sample.name,
  removestrings,
  subset,
  columns = NULL,
  notcolumns = NULL,
  subsample = NULL,
  export = FALSE,
  outpath = NULL,
  ...
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{columns}{A subset of columns to pass instead}

\item{notcolumns}{A subset of columns to remove}

\item{subsample}{If downsampling is wanted.}

\item{outpath}{Location to store new .fcs files}

\item{subsets}{The subset of interest from gating hierarchy}
}
\value{
Two columns for the Dimensionality visualized data
}
\description{
Runs PaCMAP via Basilisk
}
\examples{
NULL
}
