% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParallelNxNPlots.R
\name{Utility_ParallelNbyNPlots}
\alias{Utility_ParallelNbyNPlots}
\title{Overlay plot two .fcs files from two different gs.}
\usage{
Utility_ParallelNbyNPlots(
  x,
  y,
  sample.name,
  removestrings,
  marginsubset,
  gatesubset,
  ycolumn,
  bins,
  clearance,
  gatelines,
  reference = NULL,
  outpath,
  pdf
)
}
\arguments{
\item{x}{The first gs object}

\item{y}{The second gs object}

\item{sample.name}{The keyword under which sample names are stored}

\item{removestrings}{A list of values to remove from sample names}

\item{marginsubset}{The gs subset that defines the plot margin}

\item{gatesubset}{The gs subset of interest}

\item{ycolumn}{The desired y-column for the comparisons}

\item{bins}{Desired number of hex bins}

\item{clearance}{A multiplication factor for margin wiggle room (0.2)}

\item{gatelines}{Whether to plot .csv specified gate lines}

\item{reference}{Reference for .csv specified gate lines}

\item{outpath}{Location which to store the output}

\item{pdf}{Whether to return as a pdf}
}
\value{
Either list ggplot objects or a pdf object
}
\description{
Overlay plot two .fcs files from two different gs.
}
\examples{
Not at this time
}
