% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Patchwork.R
\name{Utility_Patchwork}
\alias{Utility_Patchwork}
\title{Generate the plots into a single pdf}
\usage{
Utility_Patchwork(
  x,
  filename,
  outfolder,
  thecolumns = 2,
  therows = 3,
  width = 7,
  height = 9,
  returntype = "pdf"
)
}
\arguments{
\item{x}{A list of plots}

\item{filename}{Name to save the .pdf as}

\item{outfolder}{Location to save the .pdf}

\item{thecolumns}{The number of columns per page}

\item{therows}{The number of rows per page}

\item{width}{Desired page width}

\item{height}{Desired page height}

\item{returntype}{Whether to return "pdf" (to desired location) or "patchwork" (to R)}
}
\value{
Some additional value to edit
}
\description{
Generate the plots into a single pdf
}
\examples{
NULL

}
