% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Patchwork.R
\name{Utility_Patchwork}
\alias{Utility_Patchwork}
\title{Generate the plots into a single pdf}
\usage{
Utility_Patchwork(
  x,
  filename,
  outfolder,
  thecolumns = 2,
  therows = 3,
  width = 7,
  height = 9,
  returntype = "pdf",
  NotListofList = TRUE,
  patches = FALSE
)
}
\arguments{
\item{x}{A list of plots}

\item{filename}{Name to save the .pdf as}

\item{outfolder}{Location to save the .pdf}

\item{thecolumns}{The number of columns per page}

\item{therows}{The number of rows per page}

\item{width}{Desired page width}

\item{height}{Desired page height}

\item{returntype}{Whether to return "pdf" (to desired location) or "patchwork" (to R)}

\item{NotListofList}{Internal default is True, don't touch.}

\item{patches}{Whether NotListofList is patchwork objects}
}
\value{
Some additional value to edit
}
\description{
Generate the plots into a single pdf
}
\examples{

library(flowCore)
library(flowWorkspace)
library(openCyto)
library(data.table)

File_Location <- system.file("extdata", package = "Luciernaga")
FCS_Files <- list.files(path = File_Location, pattern = ".fcs",
  full.names = TRUE)
UnstainedFCSFiles <- FCS_Files[grep("Unstained", FCS_Files)]
UnstainedCells <- UnstainedFCSFiles[-grep("Beads", UnstainedFCSFiles)]
MyCytoSet <- load_cytoset_from_fcs(UnstainedCells[1:3],
  truncate_max_range = FALSE,transformation = FALSE)
MyGatingSet <- GatingSet(MyCytoSet)
MyGates <- fread(file.path(path = File_Location, pattern = 'Gates.csv'))
MyGatingTemplate <- gatingTemplate(MyGates)
gt_gating(MyGatingTemplate, MyGatingSet)
removestrings <-  c("DR_", "Cells", ".fcs", "-", " ")
StorageLocation <- file.path("C:", "Users", "JohnDoe", "Desktop")

AllSpecimens <- Utility_IterativeGating(x=MyGatingSet[1:3],
 sample.name = "GUID", removestrings = removestrings,
 subset = "nonDebris", gate = "lymphocytes", xValue = "FSC-A",
  yValue = "SSC-A", bins = 270)

Patchwork <- Utility_Patchwork(AllSpecimens, "LymphocyteGates",
 outfolder=StorageLocation, thecolumns=2, therows=2,
  width = 7, height = 9, returntype="patchwork")

}
