% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Phate.R
\name{Utility_Phate}
\alias{Utility_Phate}
\title{Runs PHATE via Basilisk}
\usage{
Utility_Phate(
  x,
  sample.name,
  removestrings,
  subset,
  columns = NULL,
  notcolumns = NULL,
  subsample = NULL,
  export = FALSE,
  outpath = NULL,
  ...
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subset}{The subset of interest from gating hierarchy}

\item{columns}{A subset of columns to pass instead}

\item{notcolumns}{A subset of columns to remove}

\item{subsample}{If downsampling is wanted.}

\item{export}{When set to TRUE returns fcs files to specified outpath.}

\item{outpath}{Location to store new .fcs files}

\item{...}{Additional arguments passed to Phate for fine tuning}
}
\value{
Numbers
}
\description{
Runs PHATE via Basilisk
}
\examples{
NULL
}
