% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_RidgePlot.R
\name{Utility_RidgePlots}
\alias{Utility_RidgePlots}
\title{Draw Ridge Plots for comparison}
\usage{
Utility_RidgePlots(
  gs,
  subset,
  TheX = NULL,
  TheY,
  TheFill,
  inverse.transform = FALSE,
  outpath,
  returntype,
  therows = 2,
  thecolumns = 1,
  width = 7,
  height = 9,
  filename
)
}
\arguments{
\item{gs}{A Gating Set Object}

\item{subset}{The desired gate node, ie "lymphocytes"}

\item{TheX}{Optional, you can pass a list of column markers to plot.}

\item{TheY}{The desired Y-axis parameter (as named in pData)}

\item{TheFill}{The desired fill color (as named in pData)}

\item{inverse.transform}{Default is FALSE, reverts any transformation on
going from the GatingSet to the CytoSet object}

\item{outpath}{Desired storage location}

\item{returntype}{Whether to return a "pdf", "patchwork" or "plots"}

\item{therows}{The desired number of rows per page}

\item{thecolumns}{The desired number of columns per page}

\item{width}{The desired page width}

\item{height}{The desired page height}

\item{filename}{The file name for the new .pdf}
}
\value{
A pdf file
}
\description{
Draw Ridge Plots for comparison
}
\examples{
NULL
}
