% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_SCPatchwork.R
\name{Utility_SCPatchwork}
\alias{Utility_SCPatchwork}
\title{Forming a pdf to plot Luciernaga returns}
\usage{
Utility_SCPatchwork(x, filename, outfolder)
}
\arguments{
\item{x}{A list of plots}

\item{filename}{Name of the generated .pdf}

\item{outfolder}{Location to store the .pdf}
}
\value{
A value to be determined later
}
\description{
Forming a pdf to plot Luciernaga returns
}
\examples{
NULL
}
