% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_SingleColorQC.R
\name{Utility_SingleColorQC}
\alias{Utility_SingleColorQC}
\title{Main Luciernaga Function, normalized on individual cell level}
\usage{
Utility_SingleColorQC(
  x,
  subsets,
  sample.name,
  group.name,
  experiment = NULL,
  experiment.name = NULL,
  stats,
  Kept,
  external,
  sourcelocation,
  outpath,
  artificial,
  fcsexport,
  mainAF,
  AFOverlap,
  Beads,
  Brightness,
  Unstained
)
}
\arguments{
\item{x}{A Gating Set object (ex. gs or gs[[1]])}

\item{subsets}{The Gating Hierarchy level you will be sampling at}

\item{sample.name}{Keyword variable which samples are stored (ex. "GUID")}

\item{group.name}{Keyword variable which groups are stored (ex. "GROUPNAME")}

\item{experiment}{Provide directly experiment name (ex. "JAN2024")}

\item{experiment.name}{Keyword variable which experiment information is stored (ex. "TUBENAME")}

\item{stats}{Whether to take "mean" or "median"}

\item{Kept}{Whether "Raw" or "Normalized" values are retained in the Luciernaga object.}

\item{external}{An external autofluorescence to subtract from single colors.}

\item{sourcelocation}{Location where .fcs creation file is stored}

\item{outpath}{Location where created .fcs and .csv files are sent}

\item{artificial}{Whether an artificial 0 population should be added for a background autofluorescence stand in.}

\item{fcsexport}{Whether to export .fcs files, TRUE or FALSE}

\item{mainAF}{Main Autofluorescence Detector (ex. "V7-A")}

\item{AFOverlap}{Name of data.frame containing the Autofluorescence overlap of individual fluorophores for exclusion}

\item{Beads}{Whether the sample is Beads.}

\item{Brightness}{Whether sum of detectors should be returned.}

\item{Unstained}{Whether the sample is Unstained.}
}
\description{
Main Luciernaga Function, normalized on individual cell level
}
\examples{
NULL
}
