% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ThirdColorPlots.R
\name{Utility_ThirdColorPlots}
\alias{Utility_ThirdColorPlots}
\title{Highlight the location of a particular cell population on a given bi-exponential
axis.}
\usage{
Utility_ThirdColorPlots(
  x,
  subset,
  xaxis,
  yaxis,
  zaxis,
  splitpoint = 100,
  thecolor = "red",
  sample.name,
  removestrings,
  tilesize = 0.7,
  FactorNames = NULL,
  reference = NULL
)
}
\arguments{
\item{x}{A GatingSet Object}

\item{subset}{Desired Gate of Interest}

\item{xaxis}{X-axis Marker}

\item{yaxis}{Y-axis Marker}

\item{zaxis}{The Marker you want visible}

\item{splitpoint}{Point that denotes positive and negative for that marker.}

\item{thecolor}{What color should positive cells be}

\item{sample.name}{The keyword that determines a specimens name}

\item{removestrings}{A list of string characters to remove from the name}

\item{tilesize}{Equivalent of bin, determines the height and width of the tile dots.}

\item{FactorNames}{Provide factor names when splitpoint equals categorical}

\item{reference}{Provide Coereba reference when splitpoint equals crossreference}
}
\value{
A ggplot object with z-axis cells highlighted from background cells
}
\description{
Highlight the location of a particular cell population on a given bi-exponential
axis.
}
\examples{
NULL
}
