% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_tSNE.R
\name{Utility_tSNE}
\alias{Utility_tSNE}
\title{Runs tSNE (Rtsne implementation)}
\usage{
Utility_tSNE(
  x,
  sample.name,
  removestrings,
  subset,
  columns,
  notcolumns,
  subsample,
  export,
  outpath
)
}
\arguments{
\item{x}{A gating set object}

\item{sample.name}{Keyword for which sample name is stored}

\item{removestrings}{A list of things to remove from sample.name}

\item{subset}{The subset of interest from gating hierarchy}

\item{columns}{A subset of columns to pass instead}

\item{notcolumns}{A subset of columns to remove}

\item{subsample}{If downsampling is wanted.}

\item{outpath}{Location to store new .fcs files}
}
\value{
Numbers
}
\description{
Runs tSNE (Rtsne implementation)
}
\examples{
NULL
}
