% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_UnityPlots.R
\name{Utility_UnityPlot}
\alias{Utility_UnityPlot}
\title{Compare the plots for given parameters for all specimens in the gating set.}
\usage{
Utility_UnityPlot(
  x,
  y,
  GatingSet,
  marginsubset,
  gatesubset,
  sample.name,
  removestrings,
  clearance,
  bins,
  gatelines,
  reference,
  returntype,
  outpath
)
}
\arguments{
\item{x}{The desired x-axis parameter}

\item{y}{The desired y-axis parameter}

\item{GatingSet}{The Gating Set that contains the specimens}

\item{marginsubset}{The desired gate for the margins}

\item{gatesubset}{The desired gate of cells you want to plot}

\item{sample.name}{The keyword that designates different specimens}

\item{removestrings}{Values to remove from both the plot titles and the pdf}

\item{clearance}{A value of clearance multiplied to the margin}

\item{bins}{How many bins to class the cells into}

\item{gatelines}{Whether to plot the reference lines}

\item{reference}{A data.frame containing references}

\item{returntype}{Whether to return "pdf", "patchwork" or "plots"}

\item{outpath}{The desired location to send the assembled pdf to}
}
\value{
The ggplots for all the specimens, as well as the optional .pdf
}
\description{
Compare the plots for given parameters for all specimens in the gating set.
}
\examples{
 NULL
}
