% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_ViewSignature.R
\name{VisualizeSignatures}
\alias{VisualizeSignatures}
\title{Visualizes the provided fluorophore signatures from a template data.frame}
\usage{
VisualizeSignatures(
  x = NULL,
  columnname = "Sample",
  data,
  characterColumns,
  Normalize = TRUE,
  legend = TRUE,
  plotname = NULL,
  plotlinecolor = NULL
)
}
\arguments{
\item{x}{Default NULL, otherwise name for a value in the columnname column
that you want to individually visualize}

\item{columnname}{Default is Flurophore, specifies column name
from which x is filtered from}

\item{data}{A template data.frame containing the fluorophore signature}

\item{Normalize}{Whether to normalize the data based on peak
detector value, default is TRUE}

\item{legend}{Default TRUE, alternately removes plot legend}

\item{plotname}{Default NULL, alternately specifies the plot
title}

\item{plotlinecolor}{Default NULL, alternatively provide color
when only a single line}
}
\value{
Visualized plot
}
\description{
Visualizes the provided fluorophore signatures from a template data.frame
}
\examples{
A <- 2 + 2

}
