% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Concentration.R
\name{Wetlab_Concentration}
\alias{Wetlab_Concentration}
\title{Takes gated .fcs files and returns concentration and other info useful for Wetlab users.}
\usage{
Wetlab_Concentration(x, subset, nameKeyword, DilutionMultiplier, TotalVolume)
}
\arguments{
\item{x}{A mapped gating set object}

\item{subset}{Desired population node to derrive counts from}

\item{nameKeyword}{Keyword containing samples name (ex. "GROUPNAME" or c("GROUPNAME", "TUBENAME"))}

\item{DilutionMultiplier}{The dilution multiplier for the sample}

\item{TotalVolume}{Volume the specimen was resuspended in.}
}
\value{
A data.frame of useful information
}
\description{
Takes gated .fcs files and returns concentration and other info useful for Wetlab users.
}
\examples{

library(flowCore)
library(flowWorkspace)
library(CytoML)
library(dplyr)
library(purrr)

File_Location <- system.file("extdata", package = "Luciernaga")
WSP_File <- list.files(File_Location, pattern=".wsp", full.names = TRUE)
ws <- open_flowjo_xml(WSP_File[1])
gs <- flowjo_to_gatingset(ws, name=1, path = File_Location)
nameKeyword <- c("GROUPNAME", "TUBENAME")

TheData <- map(.x=gs, Wetlab_Concentration, subset = "CD45+",
  nameKeyword=nameKeyword, DilutionMultiplier=100, TotalVolume=1) \%>\%
   bind_rows()

}
