% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Concentration.R
\name{Wetlab_Concentration}
\alias{Wetlab_Concentration}
\title{Takes gated .fcs files and returns concentration and other info useful for Wetlab users.}
\usage{
Wetlab_Concentration(x, subset, nameKeyword, DilutionMultiplier, TotalVolume)
}
\arguments{
\item{x}{A mapped gating set object}

\item{subset}{Desired population node to derrive counts from}

\item{nameKeyword}{Keyword containing samples name (ex. "GROUPNAME" or c("GROUPNAME", "TUBENAME"))}

\item{DilutionMultiplier}{The dilution multiplier for the sample}

\item{TotalVolume}{Volume the specimen was resuspended in.}
}
\value{
A data.frame of useful information
}
\description{
Takes gated .fcs files and returns concentration and other info useful for Wetlab users.
}
\examples{
NULL
}
