% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlab_Rest.R
\name{Wetlab_Rest}
\alias{Wetlab_Rest}
\title{Takes Concentration output, returns re-suspension amounts for Wetlab users.}
\usage{
Wetlab_Rest(
  data,
  DesiredConcentration_MillionperML,
  MaxMLperTube,
  returntype,
  outpath = NULL,
  filename = "CellResuspensions"
)
}
\arguments{
\item{data}{The Wetlab_Concentration output (modified correct Total mL amounts)}

\item{DesiredConcentration_MillionperML}{Final desired concentration (ex. 1, 3, etc.)}

\item{MaxMLperTube}{The Final Volume of the rest tube}

\item{returntype}{Whether to return "data", "plot" or "both"}

\item{outpath}{File path of location to store the table}

\item{filename}{File name to store the table as}
}
\value{
A data.frame of resuspension measurements to get to the desired parameters
}
\description{
Takes Concentration output, returns re-suspension amounts for Wetlab users.
}
\examples{
NULL
}
